% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alternating_optimization.R
\name{alternating_optimization}
\alias{alternating_optimization}
\title{Title}
\usage{
alternating_optimization(x)
}
\arguments{
\item{x}{An object of class \code{ino}.}
}
\description{
Title
}
\examples{
# alternating_optimization()
}
