% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategies.R
\name{ao_initialization}
\alias{ao_initialization}
\title{Alternating optimization initialization}
\usage{
ao_initialization(
  x,
  arg = "data",
  by_col = TRUE,
  center = TRUE,
  scale = TRUE,
  ind_ign = NULL,
  initialization = random_initialization(),
  ncores = getOption("ino_ncores"),
  verbose = getOption("ino_progress")
)
}
\arguments{
\item{x}{An object of class \code{ino}.}

\item{initialization}{An object of class \code{strategy_call} which determines the initialization for
the .
The \code{strategy_call} is generated by any
of the strategy functions (any function with the name
\code{*_initialization}), when the \code{x} is unspecified.
Per default, \code{initialization = random_initialization()}, i.e. random
initialization.}

\item{ncores}{This function is parallelized, set the number of cores here.}

\item{verbose}{Set to \code{TRUE} (\code{FALSE}) to print (hide) progress.}
}
\value{
The updated \code{ino} object.
}
\description{
This function is an implementation of the alternating optimization
initialization strategy.
}
\keyword{strategy}
