% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategies.R
\name{check_failed_runs}
\alias{check_failed_runs}
\title{Check for failed runs}
\usage{
check_failed_runs(x, verbose = getOption("ino_progress"))
}
\arguments{
\item{x}{An object of class \code{ino}.}

\item{verbose}{A boolean, which indicates whether progress should be printed.
Set to \code{TRUE} (\code{FALSE}) to print (hide) progress.
The default is \code{getOption("ino_progress")}, which is set to \code{TRUE}
when the package is loaded.}
}
\value{
No return value, called for side effects.
}
\description{
This helper function checks for failed runs in an \code{ino} object.
}
