% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ino.R
\name{check_inputs}
\alias{check_inputs}
\title{Check inputs}
\usage{
check_inputs(...)
}
\arguments{
\item{...}{Named inputs to be checked.}
}
\value{
No return value, called for side effects.
}
\description{
This helper function centralizes several input checks.
}
\keyword{internal}
