% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{check_optimizer_active}
\alias{check_optimizer_active}
\title{Checks if the optimizers with given ids are active (i.e., not removed)}
\usage{
check_optimizer_active(optimizer, ids)
}
\arguments{
\item{optimizer}{TODO}

\item{ids}{TODO}
}
\value{
TODO
}
\description{
This helper function ... TODO
}
\examples{
\dontrun{
TODO
}
}
\keyword{internal}
