% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_grid.R
\name{create_grid}
\alias{create_grid}
\title{Create grid}
\usage{
create_grid(x, at)
}
\arguments{
\item{x}{An object of class \code{ino}.}

\item{at}{A list containing the initial values.}
}
\value{
A data frame.
}
\description{
This function creates a grid of all combinations of starting values,
data sets, and arguments considered for optimization.
}
\details{
See the vignette "Initialization strategies" for more details.
}
\examples{
# Add example.

}
\keyword{internal}
