% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategies.R
\name{data_subset_initialization}
\alias{data_subset_initialization}
\title{Data subset initialization}
\usage{
data_subset_initialization(
  x,
  how = "random",
  prop = 0.5,
  data_arg = "data",
  by_row = TRUE
)
}
\arguments{
\item{x}{An object of class \code{ino}.}

\item{how}{A character, specifying how to select the data subset. Can be one of
\code{"random"} (default), \code{"first"}, and \code{"kmeans"}, see the
details.}

\item{prop}{A numeric between 0 and 1, specifying the proportion of the data subset.}

\item{data_arg}{A character, the name of the data argument.}

\item{by_row}{A boolean,}
}
\value{
The updated input \code{x}.
}
\description{
This function is an implementation of the data subset initialization
strategy. See the vignette on initialization strategies for more details.
}
\details{
Missing.
}
\examples{
#data_subset_initialization()

}
\keyword{strategy}
