% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{earthquakes}
\alias{earthquakes}
\title{Earthquake data}
\format{
The data set is a \code{data.frame} with two integer columns, \code{year} for
the year and \code{obs} for the number of measured earthquakes.
}
\source{
The data was obtained from
\url{http://hmms-for-time-series.de/second/data/earthquakes.txt}
on 2022-03-25.
}
\usage{
data(earthquakes)
}
\description{
This data set includes the number of yearly measured earthquakes from 1900
to 2006.
}
\keyword{dataset}
