% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{f_ackley}
\alias{f_ackley}
\title{The Ackley function}
\usage{
f_ackley(x)
}
\arguments{
\item{x}{A numeric vector of length 2.}
}
\value{
A numeric value.
}
\description{
The Ackley function
}
\details{
The function has multiple local minima and one global minimum in the origin.
}
\examples{
ino:::f_ackley(c(0, 0))

}
\references{
https://en.wikipedia.org/wiki/Ackley_function
}
\keyword{function}
