% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{f_beale}
\alias{f_beale}
\title{The Beale function}
\usage{
f_beale(x)
}
\arguments{
\item{x}{A numeric vector of length 2.}
}
\value{
The function value at \code{x}, a single numeric value.
}
\description{
The Beale function
}
\details{
The function has multiple local minima and one global minimum in (3, 0.5).
}
\examples{
ino:::f_beale(c(3, 0.5))

}
\references{
https://en.wikipedia.org/wiki/Test_functions_for_optimization
}
\keyword{function}
