% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{f_ll_mmnl}
\alias{f_ll_mmnl}
\title{The log-likelihood function of the normally mixed multinomial logit model}
\usage{
f_ll_mmnl(theta, data, R = 100, neg = FALSE)
}
\arguments{
\item{theta}{The vector of model coefficients in order \code{c(b,o)}, see the details.}

\item{data}{The output of \code{\link{sim_mmnl}}.}

\item{R}{The number of random draws to approximate the integral for the mixed logit
choice probabilities.}

\item{neg}{Set to \code{TRUE} to return the negative log-likelihood value.}
}
\value{
The log-likelihood value.
}
\description{
The log-likelihood function of the normally mixed multinomial logit model
}
\details{
The order of \code{theta} is supposed to be \code{c(b,o)}, where
\itemize{
\item \code{b} denotes the coefficients
\item and \code{o} the lower-triangular elements of the lower-triangular
Cholesky root of \code{Omega}.
}
}
\examples{
N <- 10
J <- 3
b <- c(-2,0.5,2)
Omega <- diag(3)
data <- ino:::sim_mmnl(N, J, b, Omega, seed = 1)
true <- attr(data, "true")
ino:::f_ll_mmnl(theta = true, data = data)

}
\references{
https://en.wikipedia.org/wiki/Multinomial_logistic_regression
}
\seealso{
\code{\link[=sim_mmnl]{sim_mmnl()}}
}
\keyword{function}
