% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{f_ll_mmnp}
\alias{f_ll_mmnp}
\title{The log-likelihood function of the normally mixed multinomial probit model}
\usage{
f_ll_mmnp(theta, data, normal_cdf = mvtnorm::pmvnorm, negative = FALSE)
}
\arguments{
\item{theta}{A vector of model coefficients in the order \code{b}, \code{o}, \code{l}.}

\item{data}{The output of \code{\link{sim_mmnp}}.}

\item{normal_cdf}{A function that evaluates the CDF of a multivariate normal distribution. It
must take the arguments
\itemize{
\item \code{lower}, the vector of lower limits of length \code{n},
\item \code{upper}, the vector of upper limits of length \code{n},
\item \code{mean}, the mean vector of length \code{n},
\item \code{sigma}, the covariance matrix of dimension \code{n},
}
and return a single numeric value.}

\item{negative}{Set to \code{TRUE} to return the negative log-likelihood value.}
}
\value{
The log-likelihood value.
}
\description{
The log-likelihood function of the normally mixed multinomial probit model
}
\examples{
N <- 100
b <- c(-2,0.5,2)
Omega <- diag(3)
O <- t(chol(Omega))
o <- O[lower.tri(O, diag = TRUE)]
Sigma <- diag(3)
L <- t(chol(Sigma))
l <- L[lower.tri(L, diag = TRUE)][-1]
data <- ino:::sim_mmnp(N, b, o, l, seed = 1)
ino:::f_ll_mmnp(theta = c(b,o,l), data = data)

}
\references{
https://en.wikipedia.org/wiki/Multinomial_probit
}
\seealso{
\code{\link[=sim_mmnp]{sim_mmnp()}}
}
\keyword{function}
