% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{f_matyas}
\alias{f_matyas}
\title{The Matyas function}
\usage{
f_matyas(x)
}
\arguments{
\item{x}{A numeric vector of length 2.}
}
\value{
A numeric value.
}
\description{
The Matyas function
}
\details{
The function has multiple local minima and one global minimum in the origin.
}
\examples{
ino:::f_matyas(c(0, 0))

}
\references{
https://en.wikipedia.org/wiki/Test_functions_for_optimization
}
\keyword{function}
