% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{filter_comparable}
\alias{filter_comparable}
\title{TODO}
\usage{
filter_comparable(results, only_comparable)
}
\arguments{
\item{results}{TODO}

\item{only_comparable}{TODO}
}
\value{
TODO
}
\description{
This helper function ... TODO
}
\examples{
\dontrun{
TODO
}
}
\keyword{internal}
