% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_optimizer.R
\name{get_default_args}
\alias{get_default_args}
\title{Get default function arguments}
\usage{
get_default_args(fun, exclude)
}
\arguments{
\item{fun}{A function.}

\item{exclude}{A character vector.}
}
\value{
A named list.
}
\description{
This function extracts the default arguments of the function \code{fun},
excluding the names in \code{exclude}.
}
\keyword{internal}
