% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{grid_ino}
\alias{grid_ino}
\title{Create grid of parameter combinations}
\usage{
grid_ino(x)
}
\arguments{
\item{x}{An object of class \code{ino}.}
}
\value{
A list, where each element is a parameter set. Each parameter set contains
at least a placeholder for the target parameter, which is set to \code{NA}
and has to be filled by the initialization strategies. Additionally, each
parameter set contains further arguments for the target function if
available. In this case, the parameter names and identifier for the parameter
values are added as attributes \code{"par_name"} and \code{"par_id"} to the
parameter set.
}
\description{
This helper function creates a grid of all parameter combinations for an
\code{ino} object,
}
\keyword{specification}
