% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{helper_flatten}
\alias{helper_flatten}
\title{Helper function for list flattening}
\usage{
helper_flatten(x)
}
\arguments{
\item{x}{A \code{list}.}
}
\value{
A \code{list}.
}
\description{
This function flattens a nested \code{list}, see the details.
}
\details{
The input \code{list} \code{x} is transformed in the following ways:
\itemize{
\item If \code{x} has only one element, it becomes the new \code{x}. This is
repeated twice. E.g., \code{list(list(1))} becomes \code{1}.
\item Else, if all elements of \code{x} have only one element, these elements
become the elements of \code{x}. E.g., \code{list(list(1), list(2))}
becomes \code{list(1:2)}. This is repeated once.
\item Else, if each element of \code{x} is a \code{list} where each element is a
\code{list} with one element, the lowest-level \code{list} is replaced by
the one element.
E.g., \code{list(list(list(1), list(2)), list(list(3), list(4)))}
becomes \code{list(list(1, 2), list(3, 4))}.
\item Else, \code{x} remains unchanged.
}
}
\keyword{internal}
