% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{helper_subset}
\alias{helper_subset}
\title{Helper function for subsetting}
\usage{
helper_subset(
  argument,
  byrow = TRUE,
  how = "random",
  proportion = 0.5,
  centers = 2,
  ignore = integer()
)
}
\arguments{
\item{argument}{A \code{vector}, \code{matrix}, or \code{data.frame}.}

\item{byrow}{Only relevant if \code{argument} is a \code{matrix} or \code{data.frame}.
In that case, either \code{TRUE} to subset row-wise (default) or
\code{FALSE} to subset column-wise.}

\item{how}{A \code{character}, specifying how to subset. Can be one of:
\itemize{
\item \code{"random"} (default), subset at random
\item \code{"first"}, subset to the first elements
\item \code{"last"}, subset to the last elements
\item \code{"similar"}, subset to similar elements
\item \code{"dissimilar"}, subset to dissimilar elements
The options \code{"similar"} and \code{"dissimilar"} apply k-means
clustering via \code{\link[stats]{kmeans}} and require that
the argument \code{argument_name} is \code{numeric}.
}}

\item{proportion}{A \code{numeric} between \code{0} and \code{1}, specifying the
subset proportion.
By default, \code{proportion = 0.5}.}

\item{centers}{Only relevant, if \code{how = "(dis)similar"}.
In that case, passed to \code{\link[stats]{kmeans}}.
By default, \code{centers = 2}.}

\item{ignore}{Only relevant if \code{argument} is a \code{matrix} or \code{data.frame} and
\code{how = "(dis)similar"}.
In that case, a \code{integer} (vector) of row indices (or column indices
if \code{byrow = FALSE}) to ignore for clustering.}
}
\value{
The subsetted input \code{argument}.
}
\description{
Helper function for subsetting
}
\details{
This function is part of the \code{$argument(action = "subset")} method of a
\code{\link{Nop}} object and subsets a given argument.
}
\keyword{internal}
