% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hmm_ino}
\alias{hmm_ino}
\title{Example application of {ino} to fit an HMM}
\format{
An object of class \code{ino}.
}
\usage{
data("hmm_ino")
}
\description{
The ino object that is obtained when running the code provided in the hidden Markov model vignette.
}
\keyword{dataset}
