% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{initial_values_helper}
\alias{initial_values_helper}
\title{Helper function for initial values}
\usage{
initial_values_helper(initial, npar, check_initial, runs, optimizer_ids)
}
\arguments{
\item{initial}{Specify the initial point where the optimizer should start. Either:
\itemize{
\item the \code{character} \code{"random"} (the default) for random initial
values drawn from a standard normal distribution,
\item a \code{numeric} vector of length \code{npar}, the deterministic
starting point for the optimization,
\item a \code{list} of such vectors (in this case, \code{runs} is set to the
length of the \code{list}),
\item or a \code{function} without any arguments that returns a
\code{numeric} vector of length \code{npar}.
In all these cases, the same initial values are used across optimizers.
}

For more flexibility, a \code{funtion} input for \code{initial} can have
two arguments, where the first argument specifies the optimization run,
and the second argument specifies the optimizer, i.e.,
\code{initial <- function(run, optimizer) ...} for the
\code{run}-th optimization run and the \code{optimizer}-th optimizer
listed in the \code{$print()} output.}

\item{npar}{A \code{integer}, the length of the initial parameter vector.}

\item{check_initial}{Either \code{TRUE} (default) to check the initial values and fail on
misspecification, or \code{FALSE} to accept misspecifications.}

\item{runs}{An \code{integer}, the number of optimization runs.
If \code{initial} is a \code{list}, \code{runs} is set to
\code{length(initial)}.
By default, \code{runs = 1}.}

\item{optimizer_ids}{An \code{integer} \code{vector} with optimizer ids.}
}
\value{
A \code{list} with \code{runs} elements, each element is a \code{list}, where
the \code{i}-th element is a \code{numeric} of length \code{npar}, which is
the initial parameter vector for optimization with the optimizer with the id
\code{optimizer_ids[i]}.
}
\description{
This function is part of the \code{$optimize()} method of a
\code{\link{Nop}} object and creates a \code{list} of initial values.
}
\keyword{internal}
