% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ino_package.R
\docType{package}
\name{ino}
\alias{ino}
\title{ino: A package for initialization for numerical optimization}
\description{
This package provides tools for the analysis of the initialization of
numerical optimization.
}
\keyword{internal}
