% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ino_hmm}
\alias{ino_hmm}
\title{Example application to HMM likelihood}
\format{
A \code{\link{Nop}} object.
}
\usage{
data("ino_hmm")
}
\description{
See the vignette about the HMM likelihood for details:
\url{https://loelschlaeger.de/ino/articles/example_hmm.html}
}
\keyword{demo}
