% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ino_mixture}
\alias{ino_mixture}
\title{Example application to mixture likelihood}
\format{
A \code{\link{Nop}} object.
}
\usage{
data("ino_mixture")
}
\description{
See the introduction vignette for details:
\url{https://loelschlaeger.de/ino/articles/ino.html}
}
\keyword{demo}
