% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ino_probit}
\alias{ino_probit}
\title{Example application of to probit likelihood}
\format{
A \code{\link{Nop}} object.
}
\usage{
data("ino_probit")
}
\description{
See the vignette about the probit likelihood for details:
\url{https://loelschlaeger.de/ino/articles/example_probit.html}
}
\keyword{demo}
