% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{is_count}
\alias{is_count}
\title{Check for proper count}
\usage{
is_count(x, allow_zero = TRUE, error = TRUE)
}
\arguments{
\item{x}{Any object.}

\item{allow_zero}{Either \code{TRUE} to allow a zero value (default), or \code{FALSE} if not.}

\item{error}{In the case that \code{x} is not a proper count, either \code{TRUE} (default)
to throw an error or \code{FALSE} to return invisibly \code{FALSE}.}
}
\value{
If \code{error = TRUE}, either invisibly \code{TRUE} or an error is thrown.
If \code{error = FALSE}, invisibly \code{TRUE} or \code{FALSE}.
}
\description{
This function checks whether the input is a proper count, i.e., a single,
positive \code{integer}.
}
\keyword{checks}
