% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_name}
\alias{is_name}
\title{Check for proper name}
\usage{
is_name(x, error = TRUE)
}
\arguments{
\item{x}{Any object.}

\item{error}{In the case that \code{x} is not a proper name, either \code{TRUE} (default)
to throw an error or \code{FALSE} to return invisibly \code{FALSE}.}
}
\value{
If \code{error = TRUE}, either invisibly \code{TRUE} or an error is thrown.
If \code{error = FALSE}, invisibly \code{TRUE} or \code{FALSE}.
}
\description{
This function checks whether the input is a proper name, i.e., a single
(non-trivial) \code{character}.
}
\examples{
\dontrun{
is_name("one")
is_name(1)
is_name(LETTERS[1:2])
is_name("")
}
}
\keyword{utils}
