% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_number}
\alias{is_number}
\title{Check for proper number}
\usage{
is_number(x)
}
\arguments{
\item{x}{Any object.}
}
\value{
Either \code{TRUE} or \code{FALSE}.
}
\description{
This function checks whether the input is proper number, i.e., a single,
positive \code{integer}.
}
\examples{
\dontrun{
is_number(1)
is_number(pi)
is_number("1")
}
}
\keyword{utils}
