% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{is_number_vector}
\alias{is_number_vector}
\title{Check for proper number vector}
\usage{
is_number_vector(x, allow_na = TRUE, error = TRUE)
}
\arguments{
\item{x}{Any object.}

\item{allow_na}{Either \code{TRUE} to allow \code{NA} in \code{x} (default) or \code{FALSE}
to not allow \code{NA}.}

\item{error}{In the case that \code{x} is not a proper number vector, either \code{TRUE}
(default) to throw an error or \code{FALSE} to return invisibly \code{FALSE}.}
}
\value{
If \code{error = TRUE}, either invisibly \code{TRUE} or an error is thrown.
If \code{error = FALSE}, invisibly \code{TRUE} or \code{FALSE}.
}
\description{
This function checks whether the input is a proper number vector, i.e., a
\code{numeric}.
}
\keyword{checks}
