% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_proportion}
\alias{is_proportion}
\title{Check for proper proportion}
\usage{
is_proportion(x, error = TRUE)
}
\arguments{
\item{x}{Any object.}

\item{error}{In the case that \code{x} is not a proper proportion, either \code{TRUE}
(default) to throw an error or \code{FALSE} to return invisibly \code{FALSE}.}
}
\value{
If \code{error = TRUE}, either invisibly \code{TRUE} or an error is thrown.
If \code{error = FALSE}, invisibly \code{TRUE} or \code{FALSE}.
}
\description{
This function checks whether the input is proper proportion, i.e., a single
\code{numeric} between 0 and 1.
}
\examples{
\dontrun{
is_proportion(0.5)
is_proportion(pi)
is_proportion("1")
}
}
\keyword{utils}
