% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_time_limit}
\alias{is_time_limit}
\title{Check for proper time limit}
\usage{
is_time_limit(x, error = TRUE)
}
\arguments{
\item{x}{Any object.}

\item{error}{In the case that \code{x} is not a proper time limit, either \code{TRUE}
(default) to throw an error or \code{FALSE} to return invisibly \code{FALSE}.}
}
\value{
If \code{error = TRUE}, either invisibly \code{TRUE} or an error is thrown.
If \code{error = FALSE}, invisibly \code{TRUE} or \code{FALSE}.
}
\description{
This function checks whether the input is proper time limit, i.e., a single,
positive \code{numeric}.
}
\keyword{utils}
