% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{is_TRUE_FALSE}
\alias{is_TRUE_FALSE}
\title{Check for proper boolean}
\usage{
is_TRUE_FALSE(x, allow_na = TRUE, error = TRUE)
}
\arguments{
\item{x}{Any object.}

\item{allow_na}{Either \code{TRUE} to allow \code{NA} for \code{x} (default) or \code{FALSE}
to not allow \code{NA}.}

\item{error}{In the case that \code{x} is not a proper boolean, either \code{TRUE}
(default) to throw an error or \code{FALSE} to return invisibly \code{FALSE}.}
}
\value{
If \code{error = TRUE}, either invisibly \code{TRUE} or an error is thrown.
If \code{error = FALSE}, invisibly \code{TRUE} or \code{FALSE}.
}
\description{
This function checks whether the input is a proper boolean, i.e., either
\code{TRUE} or \code{FALSE}.
}
\keyword{checks}
