% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sets.R
\docType{data}
\name{logit}
\alias{logit}
\title{Simulated choice data from a mixed logit model}
\format{
Missing
}
\usage{
data(logit)
}
\description{
This data set includes simulated choices and corresponding choice
characteristics from a (normally) mixed logit model.
}
\details{
Missing
\preformatted{
NA
}
}
\keyword{dataset}
