% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sets.R
\docType{data}
\name{logit_data}
\alias{logit_data}
\title{Simulated choice data from a mixed logit model}
\format{
The format of the data sets is documented in \code{\link{sim_mnl}}.
}
\usage{
data(logit_data)
}
\description{
This object is a list which includes 100 data sets of simulated choices and
corresponding choice characteristics from a (normally) mixed logit model.
}
\details{
The data set was simulated via the following lines:
\preformatted{
  set.seed(1)
  logit_data <- list()
  for(i in 1:100){
    b <- rnorm(3, sd = 3)
    Omega <- RprobitB::rwishart(3,diag(3))$W
    name <- paste0("data",i)
    logit_data[[name]] <- ino:::sim_mnl(
      N = 300, J = 3, b = b, Omega = Omega, seed = i
    )
  }
}
}
\keyword{dataset}
