% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ino.R
\name{merge_ino}
\alias{merge_ino}
\title{Merge records}
\usage{
merge_ino(...)
}
\arguments{
\item{...}{Arbitrary many \code{ino} objects, which are merged into the first one.}
}
\value{
The updated \code{ino} object.
}
\description{
This function merges the records of multiple \code{ino} objects.
}
\keyword{specification}
