% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_optimizer.R
\name{new_optimizer}
\alias{new_optimizer}
\title{Construction of \code{optimizer} objects}
\usage{
new_optimizer(optimizer, ...)
}
\arguments{
\item{optimizer}{Either a character with the name of pre-defined optimizer or an object of
class \code{function}.}

\item{...}{Additional arguments to be passed to \code{optimizer}.}
}
\value{
An object of class \code{optimizer}.
}
\description{
This function constructs an object of class \code{optimizer}.
}
\keyword{internal}
