% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ino.R
\name{new_opts}
\alias{new_opts}
\title{Constructor}
\format{
A \code{opts} object is a list, where each element is a \code{optimizer}
object. See \code{\link{new_optimizer}} for the documentation of an
\code{optimizer} object.
}
\usage{
new_opts(x = list(), opt = set_optimizer_nlm())
}
\arguments{
\item{x}{A list.}

\item{opt}{The output of \code{\link{set_optimizer}}, which is an object of class
\code{optimizer}.
Per default, \code{opt = set_optimizer_nlm()}, which specifies the
\code{\link[stats]{nlm}} optimizer.
Can also be a (named) list of multiple \code{optimizer} objects.}
}
\value{
An object of class \code{opts}.
}
\description{
This function constructs an \code{opts} object.
}
\details{
The \code{opts} object specifies the numerical optimizer(s).
}
\keyword{internal}
