% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ino.R
\name{new_runs}
\alias{new_runs}
\title{Constructor}
\format{
A \code{runs} object is a list. For each recorded optimization run, it has
a list that stores
\itemize{
\item the name of the initialization strategy \code{.strategy},
\item the optimization time \code{.time} (as a \code{difftime} object),
\item the function value at the optimum \code{.optimum},
\item the identifier for the optimizer \code{.optimizer},
\item identifier for additional parameters for the target function (if any),
\item the initial parameter vector \code{.init},
\item the parameter estimate \code{.estimate},
\item and additional outputs of the optimizer (if any).
}
}
\usage{
new_runs(x = list())
}
\arguments{
\item{x}{A list.}
}
\value{
An object of class \code{runs}.
}
\description{
This function constructs a \code{runs} object.
}
\details{
The \code{runs} object contains results of the optimization runs.
}
\keyword{internal}
