% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ino.R
\name{nfails}
\alias{nfails}
\title{Number of failed optimization runs}
\usage{
nfails(x)
}
\arguments{
\item{x}{An object of class \code{ino}.}
}
\value{
An integer, the number of failed recorded optimization runs.
}
\description{
This function returns the number of failed optimization runs saved in an
\code{ino} object.
}
\keyword{specification}
