% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ino.R
\name{ngrid}
\alias{ngrid}
\title{Number of grid elements}
\usage{
ngrid(x)
}
\arguments{
\item{x}{An object of class \code{ino}.}
}
\value{
An integer, the number of grid elements.
}
\description{
This function returns the number of grid elements of an \code{ino} object.
}
\keyword{specification}
