% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ino-package.R
\docType{data}
\name{Nop_hmm}
\alias{Nop_hmm}
\title{Example application to HMM likelihood}
\format{
A \code{\link{Nop}} object.
}
\usage{
data("Nop_hmm")
}
\description{
This object is saved for reproducibility and to save computation time when
building the vignettes.

See the vignette about the HMM likelihood for details on how this object was
built: \url{https://loelschlaeger.de/ino/articles/example_hmm.html}
}
\keyword{internal}
