% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demos.R
\docType{data}
\name{Nop_probit}
\alias{Nop_probit}
\title{Example application to probit likelihood}
\format{
A \code{\link{Nop}} object.
}
\usage{
data("Nop_probit")
}
\description{
This object is saved for reproducibility and to save computation time when
building the vignettes.

See the vignette about the probit likelihood for details on how this object
was built:
\url{https://loelschlaeger.de/ino/articles/example_probit.html}
}
\keyword{demo}
