% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ino.R
\name{npar}
\alias{npar}
\title{Number of parameters}
\usage{
npar(x)
}
\arguments{
\item{x}{An object of class \code{ino}.}
}
\value{
An integer, the number of parameters.
}
\description{
This function extracts the \code{npar} element from an \code{ino} object,
which is the length of the parameter vector over which the target function is
optimized.
}
\keyword{specification}
