% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluations.R
\name{nr_optima}
\alias{nr_optima}
\title{Number of optima found}
\usage{
nr_optima(x, round = 2)
}
\arguments{
\item{x}{An object of class \code{ino}.}

\item{round}{TBA}
}
\value{
A ggplot.
}
\description{
This function provides an overview of the number of optima found with respect
to the initialization strategies.
}
\examples{
# nr_optima()

}
\keyword{evaluation}
