% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ino_old.R
\name{nruns}
\alias{nruns}
\title{Number of optimization runs}
\usage{
nruns(x)
}
\arguments{
\item{x}{An object of class \code{ino}.}
}
\value{
An integer, the number of recorded optimization runs.
}
\description{
This function returns the number of optimization runs saved in an \code{ino}
object.
}
\keyword{specification}
