% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization_time.R
\name{optimization_time}
\alias{optimization_time}
\title{Title}
\usage{
optimization_time(x)
}
\arguments{
\item{x}{An object of class \code{ino}.}
}
\description{
Title
}
\examples{
# optimization_time()
}
