% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizer.R
\name{optimizer_ao}
\alias{optimizer_ao}
\title{Specify ao optimizer}
\usage{
optimizer_ao(..., out_ign = character(), test_par = list())
}
\arguments{
\item{...}{Additional arguments to be passed to the optimizer. Without
specifications, the default values of the optimizer are used.}
}
\value{
An object of class \code{optimizer}.
}
\description{
This function is a wrapper for \code{\link[optimizeR]{set_optimizer}} with
the \code{\link[ao]{ao}} optimizer.
}
