% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluations.R
\name{overview_optima}
\alias{overview_optima}
\title{Optima overview}
\usage{
overview_optima(x, digits = 2)
}
\arguments{
\item{x}{An object of class \code{ino}.}

\item{digits}{The number of decimal places of the optima values. The default is \code{2}.}
}
\value{
A data frame with columns
\describe{
\item{optimum}{the unique optima (with respect to \code{digits})}
\item{frequency}{the number of runs the optima was reached}
}
}
\description{
This function provides an overview of the identified optima.
}
\keyword{evaluation}
