% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluations.R
\name{overview_vars}
\alias{overview_vars}
\title{Variables overview}
\usage{
overview_vars(x)
}
\arguments{
\item{x}{An object of class \code{ino}.}
}
\value{
A data frame with columns
\describe{
\item{name}{}
\item{}{}
}
}
\description{
This function provides an overview of the available optimization variables.
}
\keyword{evaluation}
