% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sets.R
\docType{data}
\name{probit_data}
\alias{probit_data}
\title{Simulated choice data from a mixed probit model}
\format{
Missing
}
\usage{
data(probit_data)
}
\description{
This data set includes simulated choices and corresponding choice
characteristics from a (normally) mixed probit model.
}
\details{
Missing
\preformatted{
NA
}
}
\keyword{dataset}
