% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_initialization.R
\name{random_initialization}
\alias{random_initialization}
\title{Title}
\usage{
random_initialization(x, runs, lower, upper, sampler = "normal")
}
\arguments{
\item{x}{An object of class \code{ino}.}

\item{runs}{The number of random initializations.}

\item{lower}{The lower limit(s) for the random initial points.}

\item{upper}{The upper limit(s) for the random initial points.}

\item{sampler}{One of \code{"normal"} and \code{"uniform"}.}
}
\description{
Title
}
\examples{
# random_initialization()
}
