% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/records.R
\name{Records}
\alias{Records}
\title{Records Object}
\value{
A \code{Records} object, which is an R6 class that stores results of
numerical optimization runs.
}
\description{
A \code{Records} object stores results of numerical optimizations runs.
This object is contained inside a \code{\link{Nop}} object.
}
\details{
Results of numerical optimization runs can be provided in three formats:
\enumerate{
\item As a \code{list} that contains the results of a single optimization with
a single optimizer. This is indicated via \code{results_depth = 1}.
Such a \code{list} typically contains the elements
\itemize{
\item \code{"value"}, the function value at the optimum,
\item \code{"parameter"}, the parameters at the optimum,
\item \code{"seconds"}, the optimization time in seconds,
\item \code{"initial"}, the initial parameter,
\item \code{"error"}, indicating whether an error occurred,
\item \code{"error_message"}, the error message (if any),
and optionally other elements that are optimizer-specific.
}
\item As a \code{list} that contains the results of a single optimization run,
where each element is a \code{list} of results from one optimizer.
This is indicated via \code{results_depth = 2}.
\item As a \code{list} that contains the results of multiple optimization runs,
where each element is a \code{list} of a single optimization run.
This is indicated via \code{results_depth = 3}.
}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{optimization_labels}}{Unique stored optimization labels.}

\item{\code{optimizer_labels}}{Unique stored optimizer labels.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Records-new}{\code{Records$new()}}
\item \href{#method-Records-save}{\code{Records$save()}}
\item \href{#method-Records-get}{\code{Records$get()}}
\item \href{#method-Records-print}{\code{Records$print()}}
\item \href{#method-Records-clear}{\code{Records$clear()}}
\item \href{#method-Records-elements}{\code{Records$elements()}}
\item \href{#method-Records-number}{\code{Records$number()}}
\item \href{#method-Records-clone}{\code{Records$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Records-new"></a>}}
\if{latex}{\out{\hypertarget{method-Records-new}{}}}
\subsection{Method \code{new()}}{
Creates a new \code{Records} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Records$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A new \code{Records} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Records-save"></a>}}
\if{latex}{\out{\hypertarget{method-Records-save}{}}}
\subsection{Method \code{save()}}{
Saves optimization results.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Records$save(
  results,
  results_depth,
  optimizer_label,
  optimization_label,
  comparable
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{results}}{A \code{list} of optimization results, see the details.}

\item{\code{results_depth}}{Either \code{1}, \code{2}, or \code{3}, indicating the format of the
input \code{results}, see the details.}

\item{\code{optimizer_label}}{A \code{character}, the optimizer label(s).}

\item{\code{optimization_label}}{A \code{character}, a label for the optimization.}

\item{\code{comparable}}{Either \code{TRUE} if the results are obtained for the original optimization
problem without any transformations, or \code{FALSE} else.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Invisibly the \code{Records} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Records-get"></a>}}
\if{latex}{\out{\hypertarget{method-Records-get}{}}}
\subsection{Method \code{get()}}{
Extract optimization results.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Records$get(
  which_run = "all",
  which_optimizer = "all",
  which_element = "all",
  only_comparable = FALSE,
  by_run = TRUE,
  by_optimizer = TRUE,
  simplify = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{which_run}}{Select saved results of optimization runs. Either:
\itemize{
\item \code{"all"} for all results,
\item \code{"last"}, the last saved results,
\item \code{"failed"}, the results from all failed optimization runs,
\item a \code{character} (vector) of optimization labels,
\item a \code{numeric} (vector) of run ids.
}}

\item{\code{which_optimizer}}{Select specified numerical optimizers. Either:
\itemize{
\item \code{"all"} for all specified optimizers,
\item a \code{character} (vector) of specified optimizer labels.
}}

\item{\code{which_element}}{Select elements of saved optimization results. Either:
\itemize{
\item \code{"all"} for all available elements,
\item \code{"default"}, the elements that are saved for all optimization runs by
default, i.e.
\itemize{
\item \code{"optimization_label"}, the label for the optimization run,
\item \code{"optimizer_label"}, the label for the optimizer,
\item \code{"value"}, the function value at the found optimum,
\item \code{"parameter"}, the parameter at which the optimum value is obtained,
\item \code{"seconds"}, the optimization time in seconds,
\item \code{"initial"}, the initial parameter,
\item \code{"error"}, indicating whether an error occurred,
\item \code{"error_message"}, the error message (if any),
\item \code{"comparable"}, indicating whether the results are obtained for the
original optimization problem without any transformations.
}
\item a \code{character} (vector) with names of specific elements (see
\code{$elements_available()} for the names of all available elements).
}}

\item{\code{only_comparable}}{Either \code{TRUE} to show only comparable results (i.e., results obtained
for the original optimization problem without any transformations),
or \code{FALSE} to include all optimization results.}

\item{\code{simplify}}{If \code{simplify = TRUE}, the nested list output of optimization results is
flattened if possible.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{list} of selected optimization results.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Records-print"></a>}}
\if{latex}{\out{\hypertarget{method-Records-print}{}}}
\subsection{Method \code{print()}}{
Prints details of the stored numerical optimization results.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Records$print(digits = getOption("digits", default = 7), ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{digits}}{An \code{integer}, the number of shown decimal places.}

\item{\code{...}}{Currently not used.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Invisibly the \code{Results} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Records-clear"></a>}}
\if{latex}{\out{\hypertarget{method-Records-clear}{}}}
\subsection{Method \code{clear()}}{
Deletes optimization results.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Records$clear(which_run, which_optimizer, which_element)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{which_run}}{Select saved results of optimization runs. Either:
\itemize{
\item \code{"all"} for all results,
\item \code{"last"}, the last saved results,
\item \code{"failed"}, the results from all failed optimization runs,
\item a \code{character} (vector) of optimization labels,
\item a \code{numeric} (vector) of run ids.
}}

\item{\code{which_optimizer}}{Select specified numerical optimizers. Either:
\itemize{
\item \code{"all"} for all specified optimizers,
\item a \code{character} (vector) of specified optimizer labels.
}}

\item{\code{which_element}}{Select elements of saved optimization results. Either:
\itemize{
\item \code{"all"} for all available elements,
\item \code{"default"}, the elements that are saved for all optimization runs by
default, i.e.
\itemize{
\item \code{"optimization_label"}, the label for the optimization run,
\item \code{"optimizer_label"}, the label for the optimizer,
\item \code{"value"}, the function value at the found optimum,
\item \code{"parameter"}, the parameter at which the optimum value is obtained,
\item \code{"seconds"}, the optimization time in seconds,
\item \code{"initial"}, the initial parameter,
\item \code{"error"}, indicating whether an error occurred,
\item \code{"error_message"}, the error message (if any),
\item \code{"comparable"}, indicating whether the results are obtained for the
original optimization problem without any transformations.
}
\item a \code{character} (vector) with names of specific elements (see
\code{$elements_available()} for the names of all available elements).
}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Invisibly the \code{Records} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Records-elements"></a>}}
\if{latex}{\out{\hypertarget{method-Records-elements}{}}}
\subsection{Method \code{elements()}}{
Returns names of available elements by optimizer in the results.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Records$elements(
  which_run = "all",
  which_optimizer = "all",
  only_comparable = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{which_run}}{Select saved results of optimization runs. Either:
\itemize{
\item \code{"all"} for all results,
\item \code{"last"}, the last saved results,
\item \code{"failed"}, the results from all failed optimization runs,
\item a \code{character} (vector) of optimization labels,
\item a \code{numeric} (vector) of run ids.
}}

\item{\code{which_optimizer}}{Select specified numerical optimizers. Either:
\itemize{
\item \code{"all"} for all specified optimizers,
\item a \code{character} (vector) of specified optimizer labels.
}}

\item{\code{only_comparable}}{Either \code{TRUE} to show only comparable results (i.e., results obtained
for the original optimization problem without any transformations),
or \code{FALSE} to include all optimization results.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{list}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Records-number"></a>}}
\if{latex}{\out{\hypertarget{method-Records-number}{}}}
\subsection{Method \code{number()}}{
Returns the number of saved optimization results.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Records$number(
  which_run = "all",
  which_optimizer = "all",
  only_comparable = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{which_run}}{Select saved results of optimization runs. Either:
\itemize{
\item \code{"all"} for all results,
\item \code{"last"}, the last saved results,
\item \code{"failed"}, the results from all failed optimization runs,
\item a \code{character} (vector) of optimization labels,
\item a \code{numeric} (vector) of run ids.
}}

\item{\code{which_optimizer}}{Select specified numerical optimizers. Either:
\itemize{
\item \code{"all"} for all specified optimizers,
\item a \code{character} (vector) of specified optimizer labels.
}}

\item{\code{only_comparable}}{Either \code{TRUE} to show only comparable results (i.e., results obtained
for the original optimization problem without any transformations),
or \code{FALSE} to include all optimization results.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An \code{integer}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Records-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Records-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Records$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
