% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{result_ino}
\alias{result_ino}
\title{Save results of optimization run}
\usage{
result_ino(x, strategy, pars, result, opt_name)
}
\arguments{
\item{x}{An object of class \code{ino}.}

\item{strategy}{A character, the name of the initialization strategy.}

\item{pars}{A list of parameter values for the optimization run.}

\item{result}{The output of \code{\link{do.call_timed}}.}

\item{opt_name}{The identifier of the optimizer.}
}
\value{
The updated object \code{x} (invisibly).
}
\description{
This helper function saves the results of an optimization run into the
submitted \code{ino} object.
}
\keyword{specification}
