% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runs.R
\name{Runs}
\alias{Runs}
\title{Runs Object (R6 Class)}
\value{
A \code{Runs} object, which is an R6 class that stores results of numerical
optimization runs.
}
\description{
A \code{Runs} object stores results of numerical optimization runs. It is
typically contained inside a \code{\link{Nop}} object.
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{optimization_labels}}{All unique optimization labels.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Runs-new}{\code{Runs$new()}}
\item \href{#method-Runs-save_results}{\code{Runs$save_results()}}
\item \href{#method-Runs-save_run}{\code{Runs$save_run()}}
\item \href{#method-Runs-save_optimization}{\code{Runs$save_optimization()}}
\item \href{#method-Runs-print}{\code{Runs$print()}}
\item \href{#method-Runs-summary}{\code{Runs$summary()}}
\item \href{#method-Runs-plot}{\code{Runs$plot()}}
\item \href{#method-Runs-clear}{\code{Runs$clear()}}
\item \href{#method-Runs-elements_available}{\code{Runs$elements_available()}}
\item \href{#method-Runs-number_results}{\code{Runs$number_results()}}
\item \href{#method-Runs-clone}{\code{Runs$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Runs-new"></a>}}
\if{latex}{\out{\hypertarget{method-Runs-new}{}}}
\subsection{Method \code{new()}}{
Creates a new \code{Runs} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Runs$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A new \code{Runs} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Runs-save_results"></a>}}
\if{latex}{\out{\hypertarget{method-Runs-save_results}{}}}
\subsection{Method \code{save_results()}}{
Saves (multiple) optimization results.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Runs$save_results(results, optimizer_label, optimization_label, comparable)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{results}}{A \code{list}, each element contains the results of one optimization run,
see the argument \code{run}.}

\item{\code{optimizer_label}}{A \code{character}, the label(s) for the optimizer(s).}

\item{\code{optimization_label}}{A \code{character}, a label for the optimization.}

\item{\code{comparable}}{Either \code{TRUE} if the optimization results are obtained
for the original optimization problem without any transformations,
or \code{FALSE} else.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Invisibly the \code{Runs} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Runs-save_run"></a>}}
\if{latex}{\out{\hypertarget{method-Runs-save_run}{}}}
\subsection{Method \code{save_run()}}{
Saves a single optimization run.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Runs$save_run(run, optimizer_label, optimization_label, comparable)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{run}}{A \code{list}, each element is a \code{list} of results from
one optimizer, see the argument \code{optimization}.}

\item{\code{optimizer_label}}{A \code{character}, the label(s) for the optimizer(s).}

\item{\code{optimization_label}}{A \code{character}, a label for the optimization.}

\item{\code{comparable}}{Either \code{TRUE} if the optimization results are obtained
for the original optimization problem without any transformations,
or \code{FALSE} else.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Invisibly the \code{Runs} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Runs-save_optimization"></a>}}
\if{latex}{\out{\hypertarget{method-Runs-save_optimization}{}}}
\subsection{Method \code{save_optimization()}}{
Saves a single optimization.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Runs$save_optimization(
  optimization,
  optimizer_label,
  optimization_label,
  comparable
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{optimization}}{A \code{list} of results from a single optimization, typically containing
\itemize{
\item \code{"value"}, the function value at the found optimum,
\item \code{"parameter"}, the parameters at the optimum,
\item \code{"seconds"}, the optimization time in seconds,
\item \code{"initial"}, the initial parameter,
\item \code{"error"}, indicating whether an error occurred,
\item \code{"error_message"}, the error message (if any),
and optionally other elements that are optimizer-specific.
}}

\item{\code{optimizer_label}}{A \code{character}, the label(s) for the optimizer(s).}

\item{\code{optimization_label}}{A \code{character}, a label for the optimization.}

\item{\code{comparable}}{Either \code{TRUE} if the optimization results are obtained
for the original optimization problem without any transformations,
or \code{FALSE} else.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Invisibly the \code{Runs} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Runs-print"></a>}}
\if{latex}{\out{\hypertarget{method-Runs-print}{}}}
\subsection{Method \code{print()}}{
Prints details of the stored numerical optimization results.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Runs$print(digits = getOption("ino_digits", default = 2), ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{digits}}{An \code{integer}, the number of shown decimal places.
The default is \code{2}.}

\item{\code{...}}{Currently not used.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Invisibly the \code{Results} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Runs-summary"></a>}}
\if{latex}{\out{\hypertarget{method-Runs-summary}{}}}
\subsection{Method \code{summary()}}{
TODO
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Runs$summary()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Runs-plot"></a>}}
\if{latex}{\out{\hypertarget{method-Runs-plot}{}}}
\subsection{Method \code{plot()}}{
TODO
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Runs$plot()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Runs-clear"></a>}}
\if{latex}{\out{\hypertarget{method-Runs-clear}{}}}
\subsection{Method \code{clear()}}{
Deletes optimization results.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Runs$clear(which_run, which_optimizer, which_element)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{which_run}}{Select saved results of optimization runs. Either:
\itemize{
\item \code{"all"} for all results,
\item \code{"last"}, the last saved results,
\item \code{"failed"}, the results from all failed optimization runs,
\item a \code{character} (vector) of optimization labels,
\item a \code{numeric} (vector) of run ids.
}}

\item{\code{which_optimizer}}{Select optimizers, a \code{character} (vector) of optimizer labels.}

\item{\code{which_element}}{Select elements of saved optimization results. Either:
\itemize{
\item \code{"all"} for all available elements,
\item \code{"default"}, the elements that are saved for all optimization runs by
default, i.e.
\itemize{
\item \code{"optimization_label"}, the label for the optimization run,
\item \code{"optimizer_label"}, the label for the optimizer,
\item \code{"value"}, the function value at the found optimum,
\item \code{"parameter"}, the parameter at which the optimum value is obtained,
\item \code{"seconds"}, the optimization time in seconds,
\item \code{"initial"}, the initial parameter,
\item \code{"error"}, indicating whether an error occurred,
\item \code{"error_message"}, the error message (if any).
}
\item a \code{character} (vector) with names of specific elements (see
\code{$elements_available()} for the names of all available elements).
}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Invisibly the \code{Runs} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Runs-elements_available"></a>}}
\if{latex}{\out{\hypertarget{method-Runs-elements_available}{}}}
\subsection{Method \code{elements_available()}}{
Returns names of available elements per optimizer.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Runs$elements_available(which_optimizer)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{which_optimizer}}{Select optimizers, a \code{character} (vector) of optimizer labels.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{list}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Runs-number_results"></a>}}
\if{latex}{\out{\hypertarget{method-Runs-number_results}{}}}
\subsection{Method \code{number_results()}}{
Returns the number of saved optimization results
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Runs$number_results(
  which_run = "all",
  which_optimizer = "all",
  only_comparable = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{which_run}}{Select saved results of optimization runs. Either:
\itemize{
\item \code{"all"} for all results,
\item \code{"last"}, the last saved results,
\item \code{"failed"}, the results from all failed optimization runs,
\item a \code{character} (vector) of optimization labels,
\item a \code{numeric} (vector) of run ids.
}}

\item{\code{which_optimizer}}{Select optimizers, a \code{character} (vector) of optimizer labels.}

\item{\code{only_comparable}}{Either \code{TRUE} to show only comparable results (i.e., results obtained
for the original optimization problem without any transformations),
or \code{FALSE} to include all optimization results.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An \code{integer}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Runs-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Runs-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Runs$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
