% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_optimization_results.R
\name{save_optimization_results}
\alias{save_optimization_results}
\title{Saving of optimization results.}
\usage{
save_optimization_results(x, strategy, res, time)
}
\arguments{
\item{x}{An object of class \code{ino}.}

\item{strategy}{A character, the label for the strategy.}

\item{res}{A list object, the optimization output.}

\item{time}{An object of class \code{difftime}, the optimization time.}
}
\value{
An object of class \code{ino}.
}
\description{
This function saves optimization results in an \code{ino} object.
}
\keyword{internal}
