% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_optimizer.R
\name{set_custom_args}
\alias{set_custom_args}
\title{Set custom function arguments}
\usage{
set_custom_args(args, custom_args)
}
\arguments{
\item{args}{A named list.}

\item{custom_args}{A named list.}
}
\value{
A named list.
}
\description{
This function sets elements in \code{args} to the custom values in
\code{custom_args}.
}
\keyword{internal}
