% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_data.R
\name{set_data}
\alias{set_data}
\title{Set data for the numerical optimization of likelihood functions.}
\usage{
set_data(x, data)
}
\arguments{
\item{x}{Either \code{NULL} or an object of class \code{ino}.}

\item{data}{A list of data sets.}
}
\value{
An object of class \code{ino}.
}
\description{
Set data for the numerical optimization of likelihood functions.
}
\keyword{specification}
