% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_f.R
\name{set_f}
\alias{set_f}
\title{Set a function for the numerical optimization.}
\usage{
set_f(f, npar, ...)
}
\arguments{
\item{f}{An object of class \code{function}.}

\item{npar}{The number of parameters in \code{f}.}

\item{...}{Additional arguments to be passed to \code{f}.}
}
\value{
An object of class \code{ino_f}.
}
\description{
Set a function for the numerical optimization.
}
\examples{
himmelblau <- function(x, c, d) (x[1]^2 + x[2] + c)^2 + (x[1] + x[2]^2 + d)^2
set_f(f = himmelblau, npar = 2, c = -11, d = -7)
}
