% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{set_optimizer_optim}
\alias{set_optimizer_optim}
\title{Specify \code{\link[stats]{optim}} optimizer}
\usage{
set_optimizer_optim(..., crit = c("convergence"))
}
\arguments{
\item{...}{Additional arguments to be passed to the optimizer \code{opt}. Without
specifications, the default values of \code{opt} are used.}

\item{crit}{The names of additional elements in the output of \code{opt} to be saved
after the optimization.}
}
\value{
An object of class \code{optimizer}, which can be passed to
\code{\link{setup_ino}}.
}
\description{
Specify \code{\link[stats]{optim}} optimizer
}
\seealso{
\code{\link[=set_optimizer]{set_optimizer()}} for specifying a different optimizer.
}
\keyword{specification}
