% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{set_true}
\alias{set_true}
\title{Set true parameter values}
\usage{
set_true()
}
\arguments{
\item{NA}{}
}
\value{
NA
}
\description{
This function specifies true parameter values.
}
\examples{
NA

}
\keyword{specification}
