% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup_ino}
\alias{setup_ino}
\title{Setup of an \code{ino} object}
\usage{
setup_ino(f, npar, ..., opt = set_optimizer("nlm"), mpvs = character(0))
}
\arguments{
\item{f}{An object of class \code{function}, the function to be optimized.}

\item{npar}{The length of the first argument of \code{f}, i.e. the argument over which
\code{f} is optimized.}

\item{...}{Additional and named arguments to be passed to \code{f} (optional).}

\item{opt}{The output of \code{set_optimizer()}.
Per default, \code{opt = set_optimizer("nlm")}.
Can also be a list, see the details.}

\item{mpvs}{A character vector of the parameter names with multiple values, see the
details. Per default, \code{mpvs = character(0)}.}
}
\value{
An object of class \code{ino}.
}
\description{
Use this function to specify the numerical optimization problem. The function
returns an object of class \code{ino} that contains all specifications.
}
\details{
\subsection{Specifying a function}{

One function \code{f} must be specified per \code{ino} object. The function
is optimized over its first argument, which should be a numeric vector of
length \code{npar}, followed by any other arguments specified via the
\code{...} argument.
}

\subsection{Specifying an optimizer}{

The numerical optimizer must be specified via the \code{opt} argument as the
output of \code{set_optimizer()}. Specifying multiple optimizer is possible,
see below.
}

\subsection{Specifying multiple parameter values}{

You can specify multiple values for each \code{...} parameter and multiple
optimizer (i.e. multiple values for \code{opt}). Such arguments
must be a list, where each list element must be a valid parameter value.
The names of these arguments must be added to the \code{mpvs} character
vector argument to make clear that you want to iterate over them.
}

\subsection{An example}{

Say that \code{f} is a likelihood function of \code{npar} parameters that has
the additional argument \code{data}, which is supposed to be a set of data
values. Say furthermore that you want to conduct a simulation experiment of
the initialization effect for \code{f} for two different data sets.
And last, say that you want to use the \code{\link[stats]{nlm}} optimizer
with a gradient tolerance of \code{gradtol = 1e-10} and save the number of
performed iterations (see the documentation of \code{\link{set_optimizer}}).
Then, specify
\preformatted{
setup_ino(
  f = f,
  npar = npar,
  data = list("data1" = <data set 1>,
              "data2" = <data set 2>),
  opt = set_optimizer_nlm(gradtol = 1e-10, crit = "iterations"),
  mpvs = c("data", "opt")
)
}
}
}
\examples{
setup_ino(
  f = ino:::f_ll_hmm,
  npar = 4,
  data = ino::earthquakes,
  N = 2,
  neg = TRUE,
  opt = set_optimizer_nlm()
)

}
\seealso{
\code{\link[=set_optimizer]{set_optimizer()}} to specify an optimizer.
\code{\link[=update_ino]{update_ino()}} to replace or add specifications.
\code{\link[=test_ino]{test_ino()}} to test the specifications.
}
\keyword{specification}
