% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sim_mmnl}
\alias{sim_mmnl}
\title{Simulate data from a normally mixed multinomial logit model}
\usage{
sim_mmnl(N, J, b, o, seed = NULL)
}
\arguments{
\item{N}{The number of observations.}

\item{J}{The number of alternatives.}

\item{b}{The vector of linear coefficients.}

\item{o}{The vector of elements of the lower Cholesky root of the covariance matrix
of the normal mixing distribution.}

\item{seed}{Set a seed for the simulation.}
}
\value{
A list with choices \code{y} and covariate matrices \code{X}.
}
\description{
Simulate data from a normally mixed multinomial logit model
}
\examples{
N <- 100
J <- 3
b <- c(-2,0.5,2)
Omega <- diag(3)
O <- t(chol(Omega))
o <- O[lower.tri(O, diag = TRUE)]
data <- sim_mmnl(N, J, b, o, seed = 1)

}
