% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sim_mmnp}
\alias{sim_mmnp}
\title{Simulate data from a normally mixed multinomial probit model}
\usage{
sim_mmnp(N, b, o, l, seed = NULL)
}
\arguments{
\item{N}{The number of observations.}

\item{b}{The vector of linear coefficients.}

\item{o}{The vector of elements of the lower Cholesky root of the covariance matrix
of the normal mixing distribution.}

\item{l}{The vector of elements of the lower Cholesky root of the error term
covariance matrix, except for the first element which is fixed to 1 for
identification.}

\item{seed}{Set a seed for the simulation.}
}
\value{
A list with choices \code{y} and covariate matrices \code{X}.
}
\description{
Simulate data from a normally mixed multinomial probit model
}
\examples{
N <- 100
b <- c(-2,0.5,2)
Omega <- diag(3)
O <- t(chol(Omega))
o <- O[lower.tri(O, diag = TRUE)]
Sigma <- diag(3)
L <- t(chol(Sigma))
l <- L[lower.tri(L, diag = TRUE)][-1]
data <- sim_mmnp(N, b, o, l, seed = 1)
}
