% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{simplify_results}
\alias{simplify_results}
\title{TODO}
\usage{
simplify_results(results, simplify)
}
\arguments{
\item{results}{TODO}

\item{simplify}{TODO}
}
\value{
TODO
}
\description{
This helper function ... TODO
}
\examples{
\dontrun{
TODO
}
}
\keyword{internal}
