% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{standardize_argument}
\alias{standardize_argument}
\title{Standardize argument}
\usage{
standardize_argument(argument, by_column, center, scale, ignore)
}
\arguments{
\item{argument}{A \code{numeric} \code{vector}, \code{matrix}, or \code{data.frame}.}

\item{by_column, center, scale, ignore}{See documentation of method \code{$standardize()} from \code{Nop} object.}
}
\value{
The standardized \code{argument}.
}
\description{
This helper function standardizes a \code{numeric} argument.
}
\keyword{internal}
