% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{standardize_helper}
\alias{standardize_helper}
\title{Helper function for standardization}
\usage{
standardize_helper(
  argument,
  by_column = TRUE,
  center = TRUE,
  scale = TRUE,
  ignore = integer(),
  jointly = list()
)
}
\arguments{
\item{argument}{A \code{numeric} \code{vector}, \code{matrix}, or \code{data.frame}.}

\item{by_column}{Only relevant if \code{argument} is a \code{matrix} or a \code{data.frame}.
In that case, either \code{TRUE} to standardize column-wise (default) or
\code{FALSE} to standardize row-wise.}

\item{center}{Set to \code{TRUE} (default) for centering, resulting in zero mean.}

\item{scale}{Set to \code{TRUE} (default) for scaling, resulting in unit variance.}

\item{ignore}{Only relevant if \code{argument} is a \code{matrix} or \code{data.frame}.
In that case, A \code{integer} (vector) of column indices (or row indices if
\code{by_column = FALSE}) to not standardize.}

\item{jointly}{Only relevant if \code{argument} is a \code{matrix} or \code{data.frame}.
In that case, a \code{list} of \code{integer} vectors with column indices
(or row indices if \code{by_column = FALSE}) to standardize jointly.}
}
\value{
The standardized input \code{argument}.
The \code{numeric} centering and scalings used (if any) are added as
attributes \code{"standardized:center"} and \code{"standardized:scale"}.
}
\description{
Helper function for standardization
}
\details{
This function is part of the \code{$standardize()} method of a
\code{\link{Nop}} object and standardizes a given argument.
}
\keyword{internal}
