% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategies.R
\name{standardize_initialization}
\alias{standardize_initialization}
\title{Standardize initialization}
\usage{
standardize_initialization(
  x,
  arg = "data",
  by_col = TRUE,
  center = TRUE,
  scale = TRUE,
  ncores = getOption("ino_ncores"),
  col_ign = NULL,
  initialization = random_initialization(),
  verbose = getOption("ino_progress")
)
}
\arguments{
\item{x}{An object of class \code{ino}.}

\item{arg}{A character, the name of the argument to be standardized.
The argument must be of class \code{matrix} or \code{data.frame}.
Per default, \code{arg = "data"}.}

\item{by_col}{A boolean, set to \code{TRUE} (the default) to standardize column-wise, set
to \code{FALSE} to standardize by rows.}

\item{center}{A boolean, passed to \code{\link[base]{scale}}.}

\item{scale}{A boolean, passed to \code{\link[base]{scale}}.}

\item{ncores}{This function is parallelized, set the number of cores here.}

\item{col_ign}{A numeric vector of column indices (or row indices if \code{by_col = FALSE})
that are ignored when standardizing.}

\item{initialization}{TBA}

\item{verbose}{TBA}
}
\value{
The updated \code{ino} object.
}
\description{
This function is an implementation of the standardize initialization strategy.
}
\details{
See the vignette on initialization strategies for more details.
}
\examples{
#standardize_initialization()

}
\keyword{strategy}
