% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{subset_argument}
\alias{subset_argument}
\title{Subset argument}
\usage{
subset_argument(
  argument,
  by_row,
  how,
  proportion,
  centers,
  ignore,
  seed = NULL
)
}
\arguments{
\item{argument}{A \code{vector}, \code{matrix}, or \code{data.frame}.
In case of \code{how = "(dis)similar"}, it must be \code{numeric}.}

\item{by_row, how, proportion, centers, ignore, seed}{See documentation of method \code{$reduce()} from \code{Nop} object.}
}
\value{
The subsetted \code{argument}.
}
\description{
This helper function subsets an argument.
}
\keyword{internal}
