% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{test_ino}
\alias{test_ino}
\title{Test of an \code{ino} object}
\usage{
test_ino(x)
}
\arguments{
\item{x}{An object of class \code{ino}.}
}
\value{
No return value, prints the test results to the console.
}
\description{
Use this function to test the specification of an \code{ino} object.
}
\keyword{specification}
