% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{test_nop}
\alias{test_nop}
\title{Test \code{Nop} object}
\usage{
test_nop(x, at, optimizer_ids, time_limit, verbose, digits)
}
\arguments{
\item{x}{A \code{Nop} object.}

\item{at, time_limit, verbose, digits}{See documentation of method \code{$test()} from \code{Nop} object.}

\item{optimizer_ids}{A \code{vector} of indices.}
}
\value{
Invisibly \code{TRUE} if the tests are successful.
}
\description{
This helper function validates the configuration of a \code{Nop} object.
}
\keyword{internal}
