% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{transfer_nop}
\alias{transfer_nop}
\title{Transfer Nop object}
\usage{
transfer_nop(
  x,
  var = data.frame(old = c(".f_name", ".arguments", ".original_arguments",
    ".true_parameter", ".true_value", ".show_minimum", ".optimizer", ".results",
    ".runs_last", ".optimization_labels"), new = c(".f_name", ".arguments",
    ".original_arguments", ".true_parameter", ".true_value", ".minimized", ".optimizer",
    ".results", ".runs_last", ".optimization_labels"))
)
}
\arguments{
\item{x}{TODO}

\item{return}{TODO}
}
\description{
Transfer Nop object
}
\keyword{internal}
