% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{transform_results}
\alias{transform_results}
\title{Transform optimization results}
\usage{
transform_results(results, which_element, only_comparable, simplify)
}
\arguments{
\item{results}{A nested \code{list} of optimization results.
Each element corresponds to one optimization run and is a \code{list}
of results for each optimizer.
The results for each optimizer is a \code{list}, the output of
\code{\link[optimizeR]{apply_optimizer}}.}

\item{which_element}{See documentation of method \code{$results()} from \code{Nop} object.}

\item{only_comparable}{See documentation of method \code{$results()} from \code{Nop} object.}

\item{simplify}{See documentation of method \code{$results()} from \code{Nop} object.}
}
\value{
A \code{list}.
}
\description{
This helper function transforms optimization results:
\itemize{
\item filter the nested \code{list},
\item simplify by flattening the nested \code{list} (if possible).
}
}
\examples{
\dontrun{
transform_results(
  results = list("run" = list(
    "optimizer1" = list(
      "value" = 1, "comparable" = FALSE
    ),
    "optimizer2" = list(
      "value" = 2, "comparable" = TRUE
     )
  )),
  which_element = "value",
  only_comparable = TRUE,
  simplify = TRUE
)
}
}
\keyword{internal}
