% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{true_ino}
\alias{true_ino}
\title{Set true parameter values}
\usage{
true_ino()
}
\arguments{
\item{NA}{}
}
\value{
NA
}
\description{
This function specifies true parameter values.
}
\examples{
NA

}
\keyword{specification}
