% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{update_ino}
\alias{update_ino}
\title{Update of an \code{ino} object}
\usage{
update_ino(x, ...)
}
\arguments{
\item{x}{An object of class \code{ino}.}

\item{...}{Any argument that could be passed to \code{\link{setup_ino}}.}
}
\value{
The updated \code{ino} object \code{x}.
}
\description{
If you mis-specified an \code{ino} object or want to add specifications, you
do not have to start from scratch - use this function instead. The function
asks question in the console on whether specification in the \code{ino}
object should be replaced, removed, or added.
}
\examples{
# to be added

}
\keyword{specification}
