% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{user_confirm}
\alias{user_confirm}
\title{Helper function for user confirmation}
\usage{
user_confirm(question = "Question?", default = FALSE)
}
\arguments{
\item{question}{A \code{character}, the binary question to ask. It should end with a
question mark.}

\item{default}{Either \code{TRUE} or \code{FALSE} (default), the default decision.}
}
\value{
Either \code{TRUE} or \code{FALSE}.
}
\description{
This function asks in an interactive question a binary question.
}
\keyword{internal}
