% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ino.R
\name{validate_ino}
\alias{validate_ino}
\title{Validator}
\usage{
validate_ino(x = new_ino(), test_par = list())
}
\arguments{
\item{x}{An object of class \code{ino}.}

\item{test_par}{A list of test parameters for an \code{ino} object:
\itemize{
\item \code{validate}, a Boolean, set to \code{TRUE} (\code{FALSE}) to (not)
validate the \code{ino} object. Per default, \code{validate = TRUE}.
\item \code{init_rest}, a list of two elements, \code{lower} and \code{upper},
with lower and upper limits, respectively, for test values for \code{f}.
Can be single values (for joint limits) or numeric vectors of length
\code{npar} (for individual limits).
Per default, \code{lower = -1} and \code{upper = 1}.
\item \code{init_digits}, the number of decimal places for the test initial
values. Per default, \code{init_digits = 2}.
\item \code{f_checks}, the number of checks for \code{f} with random input values
(that fulfill the \code{init_rest} restrictions).
Per default, \code{f_checks = 10}.
\item \code{f_check_time}, the maximum number of seconds for a single check for
\code{f}.
A check is considered to be successful, if no error occurred
within \code{f_check_time} seconds.
Per default, \code{f_check_time = 1}.
\item \code{opt_checks}, the number of checks for \code{opt} with random initial
values (that fulfill the \code{init_rest} restrictions).
Per default, \code{opt_checks = 10}.
\item \code{opt_check_time}, the maximum number of seconds for a single check for
\code{opt}.
A check is considered to be successful, if no error occurred
within \code{opt_check_time} seconds.
}}
}
\value{
The validated input \code{x}.
}
\description{
This function validates an \code{ino} object.
}
\keyword{internal}
