% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ino.R
\name{validate_opti}
\alias{validate_opti}
\title{Validator}
\usage{
validate_opti(x = new_opti())
}
\arguments{
\item{x}{An object of class \code{opti}.}
}
\value{
The validated input \code{x}.
}
\description{
This function validates an \code{opti} object.
}
\keyword{internal}
