% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_optimizer.R
\name{validate_optimizer}
\alias{validate_optimizer}
\title{Validation of \code{optimizer} objects}
\usage{
validate_optimizer(optimizer)
}
\arguments{
\item{optimizer}{An object of class \code{optimizer}.}
}
\value{
No return value. Throws an error if \code{optimizer} is not proper.
}
\description{
This function validates an object of class \code{optimizer}.
}
\keyword{internal}
