% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ino_old.R
\name{validate_opts}
\alias{validate_opts}
\title{Validator}
\usage{
validate_opts(x = new_opts())
}
\arguments{
\item{x}{An object of class \code{opts}.}
}
\value{
The validated input \code{x}.
}
\description{
This function validates an \code{opts} object.
}
\keyword{internal}
