% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ino_old.R
\name{validate_runs}
\alias{validate_runs}
\title{Validator}
\usage{
validate_runs(x = new_runs())
}
\arguments{
\item{x}{An object of class \code{runs}.}
}
\value{
The validated input \code{x}.
}
\description{
This function validates a \code{runs} object.
}
\keyword{internal}
