% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluations.R
\name{var_names}
\alias{var_names}
\title{Variable names}
\usage{
var_names(x)
}
\arguments{
\item{x}{An object of class \code{ino}.}
}
\value{
A character vector.
}
\description{
This function returns the names of the available variables in an \code{ino}
object.
}
\keyword{evaluation}
