% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategies.R
\name{ao_initialization}
\alias{ao_initialization}
\title{Alternating optimization initialization}
\usage{
ao_initialization(
  x,
  partition,
  iterations = 10L,
  tolerance = 1e-06,
  initialization = random_initialization(),
  ncores = getOption("ino_ncores"),
  verbose = getOption("ino_progress")
)
}
\arguments{
\item{x}{An object of class \code{ino}.}

\item{initialization}{An object of class \code{strategy_call} which determines the initialization.
The \code{strategy_call} can be generated by one of the strategy functions
(any function with the name \code{*_initialization}), when the \code{x}
argument is unspecified.
Per default, \code{initialization = random_initialization()}, i.e. random
initialization.}

\item{ncores}{The number of cores for parallel computation over parameters and optimizers.
The default is \code{getOption("ino_ncores")}, which is set to \code{1}
when the package is loaded.}

\item{verbose}{A boolean, which indicates whether progress should be printed.
Set to \code{TRUE} (\code{FALSE}) to print (hide) progress.
The default is \code{getOption("ino_progress")}, which is set to \code{TRUE}
when the package is loaded.}
}
\value{
The updated \code{ino} object.
}
\description{
This function is an implementation of the alternating optimization
initialization strategy.
}
\keyword{strategy}
