% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.R
\name{autoplot.Nop}
\alias{autoplot.Nop}
\alias{autoplot.Nop_optima}
\alias{autoplot.Nop_deviation}
\alias{autoplot.Nop_results}
\title{Nop plotting methods}
\usage{
\method{autoplot}{Nop}(object, xlim = NULL, xlim2 = NULL, ...)

\method{autoplot}{Nop_optima}(object, ...)

\method{autoplot}{Nop_deviation}(object, ...)

\method{autoplot}{Nop_results}(object, ...)
}
\arguments{
\item{object}{Either: # TODO}

\item{xlim, xlim2}{[\code{numeric(2)}]\cr
Ranges for the first and second parameter to plot.

If \code{NULL}, they are derived from the specified initial values in \code{object}.}

\item{...}{Other arguments passed to specific methods.}
}
\value{
A \code{ggplot} object.
}
\description{
\itemize{
\item \code{autoplot.Nop()} # TODO
\item \code{autoplot.Nop_optima()}
\item \code{autoplot.Nop_deviation()}
\item \code{autoplot.Nop_results()}
}
}
