% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{build_initial}
\alias{build_initial}
\title{Unified function call for initial parameter specifications}
\usage{
build_initial(initial, npar, fail_bad_initial = TRUE)
}
\arguments{
\item{initial, npar, fail_bad_initial}{See documentation of method \code{$optimize()} from \code{Nop} object.}
}
\value{
A \code{function} that returns a \code{numeric} of length \code{npar}.
The \code{function} has two \code{integer} arguments:
\enumerate{
\item \code{run_id}, which selects the run,
\item \code{optimizer_id}, which selects the optimizer.
}
}
\description{
This helper function turns various formats of initial parameter
specifications into a unified function call.
}
\keyword{internal}
