% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{build_initial}
\alias{build_initial}
\title{Unified function call for initial parameter specifications}
\usage{
build_initial(initial, npar)
}
\arguments{
\item{initial}{See documentation of method \code{$optimize()} from \code{Nop} object.}

\item{npar}{See documentation of method \code{$initialize()} from \code{Nop} object.}
}
\value{
A \code{function} without any arguments that returns a \code{numeric}
of length \code{npar}.
}
\description{
This helper function turns various formats of initial parameter specification
into a unified function call.
}
\examples{
\dontrun{
build_initial("random", 2)(1)
build_initial(1:3, 3)(1)
build_initial(list(1:3, 2:4), 3)(2)
build_initial(function(run) rep(run, 4), 4)(3)
}

}
\keyword{internal}
