% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategies_helper.R
\name{check_inputs}
\alias{check_inputs}
\title{Check inputs}
\usage{
check_inputs(...)
}
\arguments{
\item{...}{Named inputs to be checked.}
}
\value{
Returns \code{NULL} invisibly.
}
\description{
This helper function centralizes several input checks.
}
\keyword{internal}
