% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ino.R
\name{clear_ino}
\alias{clear_ino}
\title{Clear records}
\usage{
clear_ino(x, which)
}
\arguments{
\item{x}{An object of class \code{ino}.}

\item{which}{Either \code{"all"} to clear all records, or alternatively a
numeric vector of row numbers from \code{summary(x)}.}
}
\value{
The updated input \code{x}.
}
\description{
This function clears initialization records saved in an \code{ino} object.
}
\keyword{specification}
