% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{f_beale}
\alias{f_beale}
\title{Beale function}
\usage{
f_beale(x)
}
\arguments{
\item{x}{A \code{numeric} vector of length 2.}
}
\value{
The function value at \code{x}, a single \code{numeric}.
}
\description{
Beale function
}
\details{
The function has multiple local minima and one global minimum in
\code{(3, 0.5)}.
}
\examples{
f_beale(c(3, 0.5))

}
\references{
https://en.wikipedia.org/wiki/Test_functions_for_optimization
}
\keyword{function}
