% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{f_easom}
\alias{f_easom}
\title{Easom function}
\usage{
f_easom(x)
}
\arguments{
\item{x}{A numeric vector of length 2.}
}
\value{
The function value at \code{x}, a single numeric value.
}
\description{
Easom function
}
\details{
The function has multiple local minima and one global minimum in (pi, pi).
}
\examples{
f_easom(c(pi, pi))

}
\references{
https://en.wikipedia.org/wiki/Test_functions_for_optimization
}
\keyword{function}
\keyword{internal}
