% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{f_ll_hmm}
\alias{f_ll_hmm}
\title{The log-likelihood function of a Poisson-hidden Markov model}
\usage{
f_ll_hmm(theta, data, N = 2)
}
\arguments{
\item{theta}{A numeric vector of model parameters.}

\item{data}{A data frame that includes a time series of counts.}

\item{N}{The number of states in the hidden Markov model.}
}
\value{
A log-likelihood value.
}
\description{
The log-likelihood function of a Poisson-hidden Markov model
}
\details{
To illustrate the numerical optimisation of the log-likelihood of a Poisson-hidden Markov model,
the {ino} package uses a data set cover the number of major earthquakes (magnitude 7 or greater)
in the world from 1900 until 2006.
}
\examples{
# f_ll_hmm()

}
\references{
https://en.wikipedia.org/wiki/Hidden_Markov_model
}
\keyword{function}
