% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{f_ll_mmnl}
\alias{f_ll_mmnl}
\title{The log-likelihood function of the normally mixed multinomial logit model}
\usage{
f_ll_mmnl(theta, data, R = 100, negative = FALSE)
}
\arguments{
\item{theta}{A numeric vector of model coefficients in the order \code{b}, \code{o}.}

\item{data}{The output of \code{\link{sim_mmnl}}.}

\item{R}{The number of random draws.}

\item{negative}{Set to \code{TRUE} to return the negative log-likelihood value.}
}
\value{
The log-likelihood value.
}
\description{
The log-likelihood function of the normally mixed multinomial logit model
}
\examples{
N <- 10
J <- 3
b <- c(-2,0.5,2)
Omega <- diag(3)
O <- t(chol(Omega))
o <- O[lower.tri(O, diag = TRUE)]
data <- sim_mmnl(N, J, b, o, seed = 1)
f_ll_mmnl(theta = c(b,o), data = data)

}
\references{
https://en.wikipedia.org/wiki/Multinomial_logistic_regression
}
\seealso{
\code{\link[=sim_mmnl]{sim_mmnl()}}
}
\keyword{function}
