% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{f_ll_mmnp}
\alias{f_ll_mmnp}
\title{The log-likelihood function of the mixed probit model}
\usage{
f_ll_mmnp(theta, data)
}
\arguments{
\item{theta}{A numeric vector of model parameters.}

\item{data}{A data frame of discrete choice data.}
}
\value{
A log-likelihood value.
}
\description{
The log-likelihood function of the mixed probit model
}
\details{
The {ino} package includes a data set that was sampled from a mixed probit
model, which can be accessed via \code{data()}.
}
\examples{
b <- c(-1,1)
Omega <- diag(2)
Sigma <- diag(2)
O <- t(chol(Omega))
L <- t(chol(Sigma))
theta <- c(b, O[lower.tri(O)], L[lower.tri(L)])

}
\references{
https://en.wikipedia.org/wiki/Probit_model
}
\keyword{function}
