% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{f_ll_mmnp}
\alias{f_ll_mmnp}
\title{The log-likelihood function of the normally mixed multinomial probit model}
\usage{
f_ll_mmnp(theta, data, normal_cdf = mvtnorm::pmvnorm, neg = FALSE)
}
\arguments{
\item{theta}{The vector of model coefficients in order \code{c(b,o,l)}, see the details.}

\item{data}{The output of \code{\link{sim_mmnp}}.}

\item{normal_cdf}{A function that evaluates the CDF of a multivariate normal distribution. It
must take the arguments
\itemize{
\item \code{lower}, the vector of lower limits of length \code{n},
\item \code{upper}, the vector of upper limits of length \code{n},
\item \code{mean}, the mean vector of length \code{n},
\item \code{sigma}, the covariance matrix of dimension \code{n},
}
and return a single numeric value.}

\item{neg}{Set to \code{TRUE} to return the negative log-likelihood value.}
}
\value{
The log-likelihood value.
}
\description{
The log-likelihood function of the normally mixed multinomial probit model
}
\details{
The order of \code{theta} is supposed to be \code{c(b,o,l)}, where
\itemize{
\item \code{b} denotes the coefficients,
\item \code{o} the lower-triangular elements of the lower-triangular
Cholesky root of \code{Omega},
\item and \code{l} the lower-triangular elements of the lower-triangular
Cholesky root \code{L} of the differenced (with respect to
alternative 1) error term covariance matrix \code{Sigma}, with the
top-left element of \code{L} fixed to 1.
}
}
\examples{
N <- 100
T <- 10
b <- c(-2,0.5,2)
Omega <- diag(3)
Sigma <- diag(3)
data <- ino:::sim_mmnp(N, T, b, Omega, Sigma, seed = 1)
true <- attr(data, "true")
ino:::f_ll_mmnp(theta = true, data = data)

}
\references{
https://en.wikipedia.org/wiki/Multinomial_probit
}
\seealso{
\code{\link[=sim_mmnp]{sim_mmnp()}} for simulating a data set from a probit model.
}
\keyword{function}
