% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixed_initialization.R
\name{fixed_initialization}
\alias{fixed_initialization}
\title{Title}
\usage{
fixed_initialization(x, at)
}
\arguments{
\item{x}{An object of class \code{ino}.}

\item{at}{The initialization point.}
}
\description{
Title
}
\examples{
# fixed_initialization()
}
