% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategies.R
\name{fixed_initialization}
\alias{fixed_initialization}
\title{Fixed initialization}
\usage{
fixed_initialization(
  x,
  at,
  ncores = getOption("ino_ncores"),
  verbose = getOption("ino_progress")
)
}
\arguments{
\item{x}{An object of class \code{ino}.}

\item{at}{A vector containing the (fixed) initial values.}

\item{ncores}{This function is parallelized, set the number of cores here.}

\item{verbose}{Set to \code{TRUE} (\code{FALSE}) to print (hide) progress.}
}
\value{
The updated \code{ino} object.
}
\description{
This function is an implementation of the fixed initialization strategy.
}
\details{
See the vignette on initialization strategies for more details.
}
\examples{
#fixed_initialization()

}
\keyword{strategy}
