% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluations.R
\name{get_fails}
\alias{get_fails}
\title{Get failure messages}
\usage{
get_fails(x, runs = NULL)
}
\arguments{
\item{x}{An object of class \code{ino}.}

\item{runs}{An integer vector, specifying the optimization runs of interest. Can be
\code{NULL} (default), in which case all runs are considered.}
}
\value{
A list of failure messages for the optimization run.
}
\description{
This function extracts failure messages from an \code{ino} object.
}
\seealso{
\code{\link[=get_vars]{get_vars()}} for extracting any available variable.
}
\keyword{evaluation}
