% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluations.R
\name{get_vars}
\alias{get_vars}
\title{Get variables}
\usage{
get_vars(x, runs = NULL, vars = NULL)
}
\arguments{
\item{x}{An object of class \code{ino}.}

\item{runs}{An integer vector, specifying the optimization runs of interest. Can be
\code{NULL} (default), in which case all runs are considered.}

\item{vars}{A character vector, specifying the variables of interest. Can be
\code{NULL} (default), in which case all variables are considered.}
}
\value{
A list, each element is a list of variables of an optimization run.
}
\description{
This function extracts available variables from an \code{ino} object.
}
\keyword{evaluation}
