% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ino.R
\name{grid_ino}
\alias{grid_ino}
\title{Create grid}
\usage{
grid_ino(x)
}
\arguments{
\item{x}{An object of class \code{ino}.}

\item{initial_values}{A numeric vector of length \code{npar(x)} with initial values.}
}
\value{
An object of class \code{grid}, which is a \code{list}, where each element
contains
\itemize{
\item the \code{initial_values} vector,
\item an \code{optimizer} object,
\item and (optionally) additional parameters to the target function.
The \code{grid} object has an attribute \code{overview}, which is a data
frame where the identifiers for the optimizers and additional parameters are
saved.
}
}
\description{
This helper function creates a grid of all parameter combinations and
optimizers for an \code{ino} object.
}
\keyword{internal}
