% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ino-package.R
\docType{package}
\name{ino-package}
\alias{ino}
\alias{ino-package}
\title{ino: Initialization of Numerical Optimization}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Analysis of the initialization for numerical optimization of real-valued functions, including likelihood functions of statistical models. See \url{https://loelschlaeger.de/ino/} for more details.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://loelschlaeger.de/ino/}
  \item \url{https://github.com/loelschlaeger/ino}
  \item Report bugs at \url{https://github.com/loelschlaeger/ino/issues}
}

}
\author{
\strong{Maintainer}: Lennart Oelschläger \email{oelschlaeger.lennart@gmail.com} (\href{https://orcid.org/0000-0001-5421-9313}{ORCID})

Authors:
\itemize{
  \item Marius Ötting \email{marius.oetting@uni-bielefeld.de} (\href{https://orcid.org/0000-0002-9373-0365}{ORCID})
}

Other contributors:
\itemize{
  \item Dietmar Bauer \email{dietmar.bauer@uni-bielefeld.de} [contributor]
}

}
\keyword{internal}
