% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_count}
\alias{is_count}
\title{Check for proper count}
\usage{
is_count(x, allow_zero = FALSE, error = TRUE)
}
\arguments{
\item{x}{Any object.}

\item{allow_zero}{Either \code{TRUE} to allow a zero value, or \code{FALSE} (default) if not.}

\item{error}{In the case that \code{x} is not a proper count, either \code{TRUE} (default)
to throw an error or \code{FALSE} to return invisibly \code{FALSE}.}
}
\value{
If \code{error = TRUE}, either invisibly \code{TRUE} or an error is thrown.
If \code{error = FALSE}, invisibly \code{TRUE} or \code{FALSE}.
}
\description{
This function checks whether the input is proper count, i.e., a single,
positive \code{integer}.
}
\examples{
\dontrun{
is_count(1)
is_count(0, allow_zero = TRUE)
is_count(pi)
is_count("1")
}
}
\keyword{utils}
