% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_name}
\alias{is_name}
\title{Check for proper name}
\usage{
is_name(x)
}
\arguments{
\item{x}{Any object.}
}
\value{
Either \code{TRUE} or \code{FALSE}.
}
\description{
This function checks whether the input is a proper name, i.e., a single
(non-trivial) \code{character}.
}
\examples{
\dontrun{
is_name("one")
is_name(1)
is_name(LETTERS[1:2])
is_name("")
}
}
\keyword{utils}
