% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_name_vector}
\alias{is_name_vector}
\title{Check for proper name vector}
\usage{
is_name_vector(x, error = TRUE)
}
\arguments{
\item{x}{Any object.}

\item{error}{In the case that \code{x} is not a proper name vector, either \code{TRUE}
(default) to throw an error or \code{FALSE} to return invisibly \code{FALSE}.}
}
\value{
If \code{error = TRUE}, either invisibly \code{TRUE} or an error is thrown.
If \code{error = FALSE}, invisibly \code{TRUE} or \code{FALSE}.
}
\description{
This function checks whether the input is a proper name vector, i.e., a
\code{vector} of (non-trivial) \code{character} values.
}
\keyword{utils}
