% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{merge_ino}
\alias{merge_ino}
\title{Merge multiple \code{ino} objects}
\usage{
merge_ino(...)
}
\arguments{
\item{...}{Arbitrary many \code{ino} objects, of which the optimization results are
merged into the first object, which is then returned.}
}
\value{
The updated \code{ino} object.
}
\description{
This function merges multiple \code{ino} objects.
}
\examples{
NA

}
\keyword{specification}
