% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_ino.R
\name{new_ino}
\alias{new_ino}
\title{Constructor of \code{ino} objects.}
\usage{
new_ino()
}
\value{
An object of class \code{ino}.
}
\description{
This function constructs an object of class \code{ino}.
}
\keyword{internal}
