% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ino.R
\name{new_runs}
\alias{new_runs}
\title{Constructor}
\format{
A \code{runs} object is a list of two elements:
\itemize{
\item The \code{table} element is a data frame. It has a row for each
recorded optimization run, columns contain optimization results. It stores
only single-value results, e.g. the optimal function value and the
optimization time.
\item The \code{pars} element is a list. It has an element for each recorded
optimization run. It stores all multi-value optimization results, e.g.
the initial parameter vector and the optimal parameter vector.
}
}
\usage{
new_runs(x = list())
}
\arguments{
\item{x}{A list.}
}
\value{
An object of class \code{runs}.
}
\description{
This function constructs a \code{runs} object.
}
\details{
The \code{runs} object contains results of the optimization runs.
}
\keyword{internal}
