% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nop.R
\name{Nop}
\alias{Nop}
\title{Nop Object}
\value{
For the output of the different methods, please refer to their respective
documentation.
}
\description{
A \code{Nop} object defines a numerical optimization problem.
}
\section{Getting started}{
\subsection{Step 1: Create a new \code{Nop} object}{

Call \code{object <- Nop$new(f, npar, ...)} where
\itemize{
\item \code{f} is a single-valued function to be optimized with respect to its
first argument,
\item \code{npar} is the length of the first argument of \code{f},
\item and \code{...} are additional arguments for \code{f}. The additional
arguments can also be managed via the \code{$argument()} method.
}
}

\subsection{Step 2: Specify one or more numerical optimizers}{

Call \code{object$set_optimizer(<optimizer object>)}, where
\code{<optimizer object>} is an object of class \code{optimizer}, which can
be created with the \code{\link[optimizeR]{define_optimizer}} function from
the \{optimizeR\} package.
Two \code{optimizer} objects are already available:
\itemize{
\item \code{\link[optimizeR]{optimizer_nlm}}
\item \code{\link[optimizeR]{optimizer_optim}}
}
}

\subsection{Step 3: Initialize and optimize}{

Call one of the following methods to initialize the optimization:
\itemize{
\item \code{object$initialize_fixed()} for fixed initial values,
\item \code{object$initialize_random()} for random initial values,
\item \code{object$initialize_continue()} for initial values based on parameter
estimates from previous optimization runs.
}

The different initialization strategies are illustrated in the package
vignettes. Next, call \code{object$optimize()} for the optimization.
}
}

\section{Methods for analyzing the results}{
A \code{Nop} object provides methods for the analysis of the saved
optimization results, with filter options for optimization runs, optimizers,
and elements (i.e., parts of the optimizer outputs):
\itemize{
\item \code{$results()} returns a \code{list} of the saved optimization results,
\item \code{$summary()} summarizes the results in a \code{data.frame},
\item \code{$optima()} returns a frequency table of the identified optima,
\item \code{$plot()} visualizes the optimization times or values,
\item \code{$best()} returns the best found parameter vector or function value.
}
}

\section{Other methods and fields}{
A \code{Nop} object also provides the following methods and fields for
convenience:
\itemize{
\item \code{object$validate()} checks the configurations of a \code{Nop} object,
\item \code{object$evaluate()} evaluates the target function at some point,
\item \code{$delete()} deletes optimization results,
\item \code{$elements()} returns the names of the available elements in the
optimizer outputs,
\item \code{$number()} returns the number of saved optimization results,
\item \code{$deviation()} calculates and plots deviations of parameters to a
reference parameter (e.g., to the true parameter vector
\code{self$true_parameter}),
\item \code{$trace()} calculates and plots the trace of an optimization,
\item \code{$f_name} stores the function name,
\item \code{$npar} stores the length of the target argument,
\item \code{$true_value} stores the true function value at the optimum
(if available),
\item \code{$true_parameter} stores the true parameter vector at the optimum
(if available),
\item \code{$fresh_label} generates a new label for the optimization.
}
}

\examples{
# Optimization of the Ackley function

Nop$new(f = f_ackley, npar = 2)$   # initialize the Nop object
  set_optimizer(optimizer_nlm())$  # define the nlm optimizer
  initialize_random(runs = 100)$   # initialize 100 times randomly
  optimize(verbose = FALSE)$       # optimize
  optima()                         # return overview of optima

}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{f_name}}{A \code{character}, the name of the function \code{f}.}

\item{\code{npar}}{An \code{integer}, the length of the target argument (i.e.,
the argument over which \code{f} is optimized).}

\item{\code{true_value}}{The true \code{numeric} optimum value of \code{f}
(if available).}

\item{\code{true_parameter}}{The true optimum \code{numeric} parameter vector
of length \code{npar} (if available), i.e., the point where \code{f}
obtains its optimum.}

\item{\code{fresh_label}}{A \code{character}, an optimization label that has
not been used yet.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Nop-new}{\code{Nop$new()}}
\item \href{#method-Nop-print}{\code{Nop$print()}}
\item \href{#method-Nop-argument}{\code{Nop$argument()}}
\item \href{#method-Nop-set_optimizer}{\code{Nop$set_optimizer()}}
\item \href{#method-Nop-evaluate}{\code{Nop$evaluate()}}
\item \href{#method-Nop-initialize_fixed}{\code{Nop$initialize_fixed()}}
\item \href{#method-Nop-initialize_random}{\code{Nop$initialize_random()}}
\item \href{#method-Nop-initialize_continue}{\code{Nop$initialize_continue()}}
\item \href{#method-Nop-optimize}{\code{Nop$optimize()}}
\item \href{#method-Nop-validate}{\code{Nop$validate()}}
\item \href{#method-Nop-elements}{\code{Nop$elements()}}
\item \href{#method-Nop-number}{\code{Nop$number()}}
\item \href{#method-Nop-results}{\code{Nop$results()}}
\item \href{#method-Nop-summary}{\code{Nop$summary()}}
\item \href{#method-Nop-delete}{\code{Nop$delete()}}
\item \href{#method-Nop-optima}{\code{Nop$optima()}}
\item \href{#method-Nop-plot}{\code{Nop$plot()}}
\item \href{#method-Nop-deviation}{\code{Nop$deviation()}}
\item \href{#method-Nop-trace}{\code{Nop$trace()}}
\item \href{#method-Nop-best_value}{\code{Nop$best_value()}}
\item \href{#method-Nop-best_parameter}{\code{Nop$best_parameter()}}
\item \href{#method-Nop-clone}{\code{Nop$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Nop-new"></a>}}
\if{latex}{\out{\hypertarget{method-Nop-new}{}}}
\subsection{Method \code{new()}}{
Creates a new \code{Nop} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Nop$new(f, npar, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{f}}{The \code{function} to be optimized.
It should return a single \code{numeric} value and
is optimized over its first argument (the target argument), which should
be a \code{numeric} vector of length \code{npar}.}

\item{\code{npar}}{An \code{integer}, the length of the first argument of \code{f} (i.e.,
the first argument over which \code{f} is optimized).}

\item{\code{...}}{Optionally additional (i.e., in addition to the target argument)
arguments for \code{f}. They must be named. They can be managed via
the \code{$argument()} method.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{Nop} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Nop-print"></a>}}
\if{latex}{\out{\hypertarget{method-Nop-print}{}}}
\subsection{Method \code{print()}}{
Prints details of the numerical optimization problem.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Nop$print(digits = getOption("digits", default = 7), ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{digits}}{An \code{integer}, the number of shown decimal places.}

\item{\code{...}}{Currently not used.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Invisibly the \code{Nop} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Nop-argument"></a>}}
\if{latex}{\out{\hypertarget{method-Nop-argument}{}}}
\subsection{Method \code{argument()}}{
Manages additional (i.e., in addition to the target argument) arguments
for \code{f}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Nop$argument(action, ..., verbose = getOption("ino_verbose", default = FALSE))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{action}}{One of:
\itemize{
\item \code{"set"} to set an argument,
\item \code{"get"} to extract an argument value,
\item \code{"remove"} to remove an argument,
\item \code{"reset"} to reset an argument to the original value,
\item \code{"modify"} to modify an argument to a new value (the original
value is saved and can be recovered via \code{"reset"}),
\item \code{"subset"} to subset an argument,
\item \code{"standardize"} to standardize a \code{numeric} argument.
}}

\item{\code{...}}{Additional parameters depending on the \code{action}:
\itemize{
\item if \code{action = "set"} or \code{"modify"}, one or more named
arguments,
\item if \code{action = "get"}, \code{"remove"}, or \code{"reset"},
the argument \code{name},
\item if \code{action = "subset"},
\itemize{
\item the argument \code{name},
\item \code{byrow}, either \code{TRUE} to subset
row-wise (default) or \code{FALSE} to subset column-wise,
\item \code{how}, a \code{character}, specifying how to subset, either
\itemize{
\item \code{"random"} (default), subset at random,
\item \code{"first"}, subset to the first elements,
\item \code{"last"}, subset to the last elements,
\item \code{"similar"}, subset to similar elements,
\item \code{"dissimilar"}, subset to dissimilar elements,
(the options \code{"similar"} and \code{"dissimilar"} apply k-means
clustering via \code{\link[stats]{kmeans}} and require that
the argument is \code{numeric}),
}
\item \code{proportion}, a \code{numeric} between \code{0} and \code{1},
specifying the subset proportion (the default is \code{0.5}),
\item \code{centers}, passed on to \code{\link[stats]{kmeans}}
if \code{how = "(dis)similar"} (by default, \code{centers = 2}),
\item \code{ignore}, an \code{integer} vector of row indices (or column
indices if \code{byrow = FALSE}) to ignore for clustering if
\code{how = "(dis)similar"},
\item \code{seed}, an \code{integer} for reproducibility,
}
\item if \code{action = "standardize"},
\itemize{
\item the argument \code{name},
\item \code{byrow}, either \code{TRUE} to standardize
row-wise or \code{FALSE} to standardize column-wise (default),
\item \code{center}, set to \code{TRUE} (default) for centering, resulting
in zero mean,
\item \code{scale}, set to \code{TRUE} (default) for scaling, resulting in
unit variance,
\item \code{ignore}, an \code{integer} vector of column indices (or row
indices if \code{byrow = TRUE}) to not standardize,
\item \code{jointly}, a \code{list} of \code{integer} vectors with column
indices (or row indices if \code{byrow = TRUE}) to standardize
jointly.
}
}}

\item{\code{verbose}}{Either \code{TRUE} to print progress and details, or \code{FALSE} to hide
such messages.
This can be defined globally via \code{options("ino_verbose" = TRUE)}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The argument value if \code{action = "get"} and invisibly the \code{Nop}
object, else. If \code{action = "standardize"}, the \code{numeric}
centering and scaling used (if any) are added as
attributes \code{"standardized:center"} and \code{"standardized:scale"}
to the argument.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Nop-set_optimizer"></a>}}
\if{latex}{\out{\hypertarget{method-Nop-set_optimizer}{}}}
\subsection{Method \code{set_optimizer()}}{
Specifies a numerical optimizer.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Nop$set_optimizer(optimizer, optimizer_label = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{optimizer}}{An object of class \code{optimizer}, which can be created via
\code{\link[optimizeR]{define_optimizer}}.}

\item{\code{optimizer_label}}{A \code{character}, a unique label for the optimizer.
By default \code{label = NULL}, in which case the default label saved
inside \code{optimizer} is used.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Invisibly the \code{Nop} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Nop-evaluate"></a>}}
\if{latex}{\out{\hypertarget{method-Nop-evaluate}{}}}
\subsection{Method \code{evaluate()}}{
Evaluates the function.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Nop$evaluate(
  at = stats::rnorm(self$npar),
  time_limit = NULL,
  hide_warnings = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{at}}{A \code{numeric} vector of length \code{npar}, the point where the
function is evaluated.
By default, \code{at = rnorm(self$npar)}, i.e., random values drawn
from a standard normal distribution.}

\item{\code{time_limit}}{An \code{integer}, a time limit in seconds. Computations are interrupted
prematurely if \code{time_limit} is exceeded.
This currently only works reliably under Windows OS.
No time limit if \code{time_limit = NULL} (the default).}

\item{\code{hide_warnings}}{Either \code{TRUE} or \code{FALSE} to hide (show) warning messages
during the function evaluation or optimization.
If \code{hide_warnings = FALSE}, warning messages are printed even if
\code{verbose = FALSE}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Either:
\itemize{
\item a \code{numeric} value, the function value at \code{at},
\item \code{"time limit reached"} if the time limit was reached,
\item the error message if the evaluation failed.
}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Nop-initialize_fixed"></a>}}
\if{latex}{\out{\hypertarget{method-Nop-initialize_fixed}{}}}
\subsection{Method \code{initialize_fixed()}}{
Defines fixed initial values for the optimization.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Nop$initialize_fixed(at, verbose = getOption("ino_verbose", default = FALSE))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{at}}{A \code{numeric} vector of length \code{npar}, the initial parameter
vector. It can also be a \code{list} of such vectors.}

\item{\code{verbose}}{Either \code{TRUE} to print progress and details, or \code{FALSE} to hide
such messages.
This can be defined globally via \code{options("ino_verbose" = TRUE)}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Invisibly the \code{Nop} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Nop-initialize_random"></a>}}
\if{latex}{\out{\hypertarget{method-Nop-initialize_random}{}}}
\subsection{Method \code{initialize_random()}}{
Defines random initial values for the optimization.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Nop$initialize_random(
  sampler = function() stats::rnorm(self$npar),
  runs = 1,
  seed = NULL,
  verbose = getOption("ino_verbose", default = FALSE)
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sampler}}{A \code{function} without any arguments that returns a \code{numeric}
vector of length \code{npar}.
By default, \code{sampler = rnorm(self$npar)}, i.e., random values drawn
from a standard normal distribution.}

\item{\code{runs}}{An \code{integer}, the number of optimization runs.
By default, \code{runs = 1}.}

\item{\code{seed}}{An \code{integer}, passed on to \code{\link{set.seed}} for reproducibility.
Can be \code{NULL} for no seed, which is the default.}

\item{\code{verbose}}{Either \code{TRUE} to print progress and details, or \code{FALSE} to hide
such messages.
This can be defined globally via \code{options("ino_verbose" = TRUE)}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Invisibly the \code{Nop} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Nop-initialize_continue"></a>}}
\if{latex}{\out{\hypertarget{method-Nop-initialize_continue}{}}}
\subsection{Method \code{initialize_continue()}}{
Defines random initial values based on previous optimization runs.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Nop$initialize_continue(
  which_run,
  which_optimizer,
  transform = function(x) x,
  verbose = getOption("ino_verbose", default = FALSE)
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{which_run}}{Selects saved results of optimization runs. Either:
\itemize{
\item \code{"all"} for all results,
\item \code{"last"} for the results of the last \code{$optimize()} call,
\item \code{"failed"}, the results from failed optimization runs,
\item a \code{character} (vector) of optimization labels,
\item an \code{integer} (vector) of optimization run ids.
}}

\item{\code{which_optimizer}}{Selects specified numerical optimizers. Either:
\itemize{
\item \code{"all"} for all specified optimizers,
\item a \code{character} (vector) of specified optimizer labels,
\item a \code{numeric} (vector) of optimizer ids, see the \code{$print()} output.
}}

\item{\code{transform}}{A \code{function} .}

\item{\code{verbose}}{Either \code{TRUE} to print progress and details, or \code{FALSE} to hide
such messages.
This can be defined globally via \code{options("ino_verbose" = TRUE)}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Invisibly the \code{Nop} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Nop-optimize"></a>}}
\if{latex}{\out{\hypertarget{method-Nop-optimize}{}}}
\subsection{Method \code{optimize()}}{
Optimizes the function.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Nop$optimize(
  which_optimizer = "all",
  seed = NULL,
  return_results = FALSE,
  save_results = TRUE,
  optimization_label = self$fresh_label,
  ncores = 1,
  verbose = getOption("ino_verbose", default = FALSE),
  simplify = FALSE,
  time_limit = NULL,
  hide_warnings = TRUE,
  reset_initial = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{which_optimizer}}{Selects specified numerical optimizers. Either:
\itemize{
\item \code{"all"} for all specified optimizers,
\item a \code{character} (vector) of specified optimizer labels,
\item a \code{numeric} (vector) of optimizer ids, see the \code{$print()} output.
}}

\item{\code{seed}}{An \code{integer}, passed on to \code{\link{set.seed}} for reproducibility.
Can be \code{NULL} for no seed, which is the default.}

\item{\code{return_results}}{Set to \code{TRUE} to return the optimization results as a \code{list}.
By default \code{return_results = FALSE}.}

\item{\code{save_results}}{Set to \code{TRUE} to save the optimization results inside the \code{Nop}
object. These results can be analyzed via different methods, see the details.
By default, \code{save_results = TRUE}.}

\item{\code{optimization_label}}{Only relevant if \code{save_results = TRUE}.
In this case, a \code{character} to specify a label for the optimization.
Labels are useful to distinguish initialization strategies.
By default, \code{optimization_label = self$fresh_label} creates a new
label.}

\item{\code{ncores}}{An \code{integer}, setting the number of CPU cores for parallel computation.
The default is \code{1}.
You can use \code{parallel::detectCores()} to detect the number of available
CPU cores.}

\item{\code{verbose}}{Either \code{TRUE} to print progress and details, or \code{FALSE} to hide
such messages.
This can be defined globally via \code{options("ino_verbose" = TRUE)}.}

\item{\code{simplify}}{If \code{simplify = TRUE}, the nested list output of optimization results is
flattened if possible.}

\item{\code{time_limit}}{An \code{integer}, a time limit in seconds. Computations are interrupted
prematurely if \code{time_limit} is exceeded.
This currently only works reliably under Windows OS.
No time limit if \code{time_limit = NULL} (the default).}

\item{\code{hide_warnings}}{Either \code{TRUE} or \code{FALSE} to hide (show) warning messages
during the function evaluation or optimization.
If \code{hide_warnings = FALSE}, warning messages are printed even if
\code{verbose = FALSE}.}

\item{\code{reset_initial}}{Either \code{TRUE} (default) to reset the initial values after the
optimization, or \code{FALSE}, else.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The return value depends on the value of \code{return_results}:
\itemize{
\item if \code{return_results = FALSE}, invisibly the \code{Nop} object,
\item if \code{return_results = TRUE}, a nested \code{list} of
optimization results. Each element corresponds to one optimization run
and is a \code{list} of results for each optimizer. The results for
each optimizer is a \code{list}, the output of
\code{\link[optimizeR]{apply_optimizer}}. If \code{simplify = TRUE},
the output is flattened if possible.
}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Nop-validate"></a>}}
\if{latex}{\out{\hypertarget{method-Nop-validate}{}}}
\subsection{Method \code{validate()}}{
Validates the configuration of a \code{Nop} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Nop$validate(
  at = stats::rnorm(self$npar),
  which_optimizer = "all",
  time_limit = 10,
  verbose = getOption("ino_verbose", default = FALSE),
  digits = getOption("digits", default = 7)
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{at}}{A \code{numeric} vector of length \code{npar}, the point where the
function and the optimizers are tested.
By default, \code{at = rnorm(self$npar)}, i.e., random values drawn
from a standard normal distribution.}

\item{\code{which_optimizer}}{Selects specified numerical optimizers. Either:
\itemize{
\item \code{"all"} for all specified optimizers,
\item a \code{character} (vector) of specified optimizer labels,
\item a \code{numeric} (vector) of optimizer ids, see the \code{$print()} output.
}}

\item{\code{time_limit}}{An \code{integer}, a time limit in seconds. Computations are interrupted
prematurely if \code{time_limit} is exceeded.
This currently only works reliably under Windows OS.
No time limit if \code{time_limit = NULL} (the default).}

\item{\code{verbose}}{Either \code{TRUE} to print progress and details, or \code{FALSE} to hide
such messages.
This can be defined globally via \code{options("ino_verbose" = TRUE)}.}

\item{\code{digits}}{An \code{integer}, the number of shown decimal places.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Invisibly \code{TRUE} if the tests are successful.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Nop-elements"></a>}}
\if{latex}{\out{\hypertarget{method-Nop-elements}{}}}
\subsection{Method \code{elements()}}{
Returns the names of all available elements by optimizers.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Nop$elements(which_optimizer = "all")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{which_optimizer}}{Selects specified numerical optimizers. Either:
\itemize{
\item \code{"all"} for all specified optimizers,
\item a \code{character} (vector) of specified optimizer labels,
\item a \code{numeric} (vector) of optimizer ids, see the \code{$print()} output.
}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{list}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Nop-number"></a>}}
\if{latex}{\out{\hypertarget{method-Nop-number}{}}}
\subsection{Method \code{number()}}{
Returns the number of saved optimization results.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Nop$number(
  which_run = "all",
  which_optimizer = "all",
  which_element = "all",
  only_comparable = FALSE,
  verbose = getOption("ino_verbose", default = FALSE)
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{which_run}}{Selects saved results of optimization runs. Either:
\itemize{
\item \code{"all"} for all results,
\item \code{"last"} for the results of the last \code{$optimize()} call,
\item \code{"failed"}, the results from failed optimization runs,
\item a \code{character} (vector) of optimization labels,
\item an \code{integer} (vector) of optimization run ids.
}}

\item{\code{which_optimizer}}{Selects specified numerical optimizers. Either:
\itemize{
\item \code{"all"} for all specified optimizers,
\item a \code{character} (vector) of specified optimizer labels,
\item a \code{numeric} (vector) of optimizer ids, see the \code{$print()} output.
}}

\item{\code{which_element}}{Selects elements of saved optimization results. Either:
\itemize{
\item \code{"all"} for all available elements,
\item \code{"default"}, the elements that are saved for all optimization runs by
default, i.e.
\itemize{
\item \code{"value"}, the function value at the found optimum,
\item \code{"parameter"}, the parameter at which the optimum value is obtained,
\item \code{"seconds"}, the optimization time in seconds,
\item \code{"initial"}, the initial parameter,
\item \code{"error"}, indicating whether an error occurred,
\item \code{"error_message"}, the error message (if any).
}
\item a \code{character} (vector) with names of specific elements (see
\code{$elements()} for the names of available elements).
}}

\item{\code{only_comparable}}{Either \code{TRUE} to show only comparable results (i.e., results obtained
for the original optimization problem without any transformations),
or \code{FALSE} to include all optimization results.}

\item{\code{verbose}}{Either \code{TRUE} to print progress and details, or \code{FALSE} to hide
such messages.
This can be defined globally via \code{options("ino_verbose" = TRUE)}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An \code{integer}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Nop-results"></a>}}
\if{latex}{\out{\hypertarget{method-Nop-results}{}}}
\subsection{Method \code{results()}}{
Returns saved optimization results.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Nop$results(
  which_run = "all",
  which_optimizer = "all",
  which_element = "all",
  only_comparable = FALSE,
  simplify = TRUE,
  add_identifier = character()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{which_run}}{Selects saved results of optimization runs. Either:
\itemize{
\item \code{"all"} for all results,
\item \code{"last"} for the results of the last \code{$optimize()} call,
\item \code{"failed"}, the results from failed optimization runs,
\item a \code{character} (vector) of optimization labels,
\item an \code{integer} (vector) of optimization run ids.
}}

\item{\code{which_optimizer}}{Selects specified numerical optimizers. Either:
\itemize{
\item \code{"all"} for all specified optimizers,
\item a \code{character} (vector) of specified optimizer labels,
\item a \code{numeric} (vector) of optimizer ids, see the \code{$print()} output.
}}

\item{\code{which_element}}{Selects elements of saved optimization results. Either:
\itemize{
\item \code{"all"} for all available elements,
\item \code{"default"}, the elements that are saved for all optimization runs by
default, i.e.
\itemize{
\item \code{"value"}, the function value at the found optimum,
\item \code{"parameter"}, the parameter at which the optimum value is obtained,
\item \code{"seconds"}, the optimization time in seconds,
\item \code{"initial"}, the initial parameter,
\item \code{"error"}, indicating whether an error occurred,
\item \code{"error_message"}, the error message (if any).
}
\item a \code{character} (vector) with names of specific elements (see
\code{$elements()} for the names of available elements).
}}

\item{\code{only_comparable}}{Either \code{TRUE} to show only comparable results (i.e., results obtained
for the original optimization problem without any transformations),
or \code{FALSE} to include all optimization results.}

\item{\code{simplify}}{If \code{simplify = TRUE}, the nested list output of optimization results is
flattened if possible.}

\item{\code{add_identifier}}{TODO}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A nested \code{list} of optimization results.
Each element corresponds to one optimization run and is a \code{list}
of results for each optimizer.
If \code{simplify = TRUE}, the output is flattened if possible.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Nop-summary"></a>}}
\if{latex}{\out{\hypertarget{method-Nop-summary}{}}}
\subsection{Method \code{summary()}}{
Provides an overview of the optimization runs.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Nop$summary(
  which_element = c("value", "parameter"),
  which_run = "all",
  which_optimizer = "all",
  digits = getOption("digits", default = 7),
  only_comparable = FALSE,
  add_identifier = c("optimization_id", "optimizer_label", "optimization_label")
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{which_element}}{Selects elements of saved optimization results. Either:
\itemize{
\item \code{"all"} for all available elements,
\item \code{"default"}, the elements that are saved for all optimization runs by
default, i.e.
\itemize{
\item \code{"value"}, the function value at the found optimum,
\item \code{"parameter"}, the parameter at which the optimum value is obtained,
\item \code{"seconds"}, the optimization time in seconds,
\item \code{"initial"}, the initial parameter,
\item \code{"error"}, indicating whether an error occurred,
\item \code{"error_message"}, the error message (if any).
}
\item a \code{character} (vector) with names of specific elements (see
\code{$elements()} for the names of available elements).
}}

\item{\code{which_run}}{Selects saved results of optimization runs. Either:
\itemize{
\item \code{"all"} for all results,
\item \code{"last"} for the results of the last \code{$optimize()} call,
\item \code{"failed"}, the results from failed optimization runs,
\item a \code{character} (vector) of optimization labels,
\item an \code{integer} (vector) of optimization run ids.
}}

\item{\code{which_optimizer}}{Selects specified numerical optimizers. Either:
\itemize{
\item \code{"all"} for all specified optimizers,
\item a \code{character} (vector) of specified optimizer labels,
\item a \code{numeric} (vector) of optimizer ids, see the \code{$print()} output.
}}

\item{\code{digits}}{An \code{integer}, the number of shown decimal places.}

\item{\code{only_comparable}}{Either \code{TRUE} to show only comparable results (i.e., results obtained
for the original optimization problem without any transformations),
or \code{FALSE} to include all optimization results.}

\item{\code{add_identifier}}{TODO}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{data.frame} with optimization results.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Nop-delete"></a>}}
\if{latex}{\out{\hypertarget{method-Nop-delete}{}}}
\subsection{Method \code{delete()}}{
Deletes optimization results.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Nop$delete(
  which_run,
  which_optimizer = "all",
  prompt = interactive(),
  replace = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{which_run}}{Selects saved results of optimization runs. Either:
\itemize{
\item \code{"all"} for all results,
\item \code{"last"} for the results of the last \code{$optimize()} call,
\item \code{"failed"}, the results from failed optimization runs,
\item a \code{character} (vector) of optimization labels,
\item an \code{integer} (vector) of optimization run ids.
}}

\item{\code{which_optimizer}}{Selects specified numerical optimizers. Either:
\itemize{
\item \code{"all"} for all specified optimizers,
\item a \code{character} (vector) of specified optimizer labels,
\item a \code{numeric} (vector) of optimizer ids, see the \code{$print()} output.
}}

\item{\code{prompt}}{TODO}

\item{\code{replace}}{TODO}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Invisibly the \code{Records} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Nop-optima"></a>}}
\if{latex}{\out{\hypertarget{method-Nop-optima}{}}}
\subsection{Method \code{optima()}}{
Provides an overview of the identified optimum values.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Nop$optima(
  digits = getOption("digits", default = 7),
  sort_by = "frequency",
  which_run = "all",
  which_optimizer = "all",
  only_comparable = TRUE,
  decreasing = (sort_by == "frequency"),
  max = getOption("max.print", default = 10)
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{digits}}{An \code{integer}, the number of shown decimal places.}

\item{\code{sort_by}}{Either:
\itemize{
\item \code{"frequency"} (default) to sort rows by frequency,
\item \code{"value"} to sort by value.
}}

\item{\code{which_run}}{Selects saved results of optimization runs. Either:
\itemize{
\item \code{"all"} for all results,
\item \code{"last"} for the results of the last \code{$optimize()} call,
\item \code{"failed"}, the results from failed optimization runs,
\item a \code{character} (vector) of optimization labels,
\item an \code{integer} (vector) of optimization run ids.
}}

\item{\code{which_optimizer}}{Selects specified numerical optimizers. Either:
\itemize{
\item \code{"all"} for all specified optimizers,
\item a \code{character} (vector) of specified optimizer labels,
\item a \code{numeric} (vector) of optimizer ids, see the \code{$print()} output.
}}

\item{\code{only_comparable}}{Either \code{TRUE} to show only comparable results (i.e., results obtained
for the original optimization problem without any transformations),
or \code{FALSE} to include all optimization results.}

\item{\code{decreasing}}{TODO}

\item{\code{max}}{An \code{integer}, specifying the maximal number of entries to be printed.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{data.frame}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Nop-plot"></a>}}
\if{latex}{\out{\hypertarget{method-Nop-plot}{}}}
\subsection{Method \code{plot()}}{
Visualizes the optimization time or value.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Nop$plot(
  which_element = "seconds",
  group_by = NULL,
  relative = FALSE,
  which_run = "all",
  which_optimizer = "all",
  only_comparable = FALSE,
  title = paste("Optimization of", self$f_name),
  xlim = c(NA, NA)
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{which_element}}{Either:
\itemize{
\item \code{"seconds"} to plot the optimization times (default)
\item \code{"value"} to plot the optimization values
}}

\item{\code{group_by}}{Either:
\itemize{
\item \code{"optimization_label"} to group by optimization label
\item \code{"optimizer_label"} to group by optimizer label
\item \code{NULL} to not group (default)
}}

\item{\code{relative}}{Only if \code{which_element = "seconds"}.
In this case, set to \code{TRUE} to plot relative time differences with
respect to the overall median.}

\item{\code{which_run}}{Selects saved results of optimization runs. Either:
\itemize{
\item \code{"all"} for all results,
\item \code{"last"} for the results of the last \code{$optimize()} call,
\item \code{"failed"}, the results from failed optimization runs,
\item a \code{character} (vector) of optimization labels,
\item an \code{integer} (vector) of optimization run ids.
}}

\item{\code{which_optimizer}}{Selects specified numerical optimizers. Either:
\itemize{
\item \code{"all"} for all specified optimizers,
\item a \code{character} (vector) of specified optimizer labels,
\item a \code{numeric} (vector) of optimizer ids, see the \code{$print()} output.
}}

\item{\code{only_comparable}}{Either \code{TRUE} to show only comparable results (i.e., results obtained
for the original optimization problem without any transformations),
or \code{FALSE} to include all optimization results.}

\item{\code{title}}{A \code{character}, the plot title.}

\item{\code{xlim}}{Passed on to \code{\link[ggplot2]{coord_cartesian}}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{\link[ggplot2]{ggplot}} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Nop-deviation"></a>}}
\if{latex}{\out{\hypertarget{method-Nop-deviation}{}}}
\subsection{Method \code{deviation()}}{
Visualizes deviation of parameters.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Nop$deviation(
  reference = self$true_parameter,
  which_element = "initial",
  which_run = "all",
  which_optimizer = "all",
  only_comparable = FALSE,
  title = "Parameter deviation",
  ylim = c(NA, NA),
  parameter_labels = paste0("theta", 1:self$npar)
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{reference}}{A \code{numeric} of length \code{self$npar}, the reference parameters.}

\item{\code{which_element}}{Either:
\itemize{
\item \code{"initial"} to compute deviations to the initial values (default)
\item \code{"parameter"} to compute deviations to the estimated parameters
}}

\item{\code{which_run}}{Selects saved results of optimization runs. Either:
\itemize{
\item \code{"all"} for all results,
\item \code{"last"} for the results of the last \code{$optimize()} call,
\item \code{"failed"}, the results from failed optimization runs,
\item a \code{character} (vector) of optimization labels,
\item an \code{integer} (vector) of optimization run ids.
}}

\item{\code{which_optimizer}}{Selects specified numerical optimizers. Either:
\itemize{
\item \code{"all"} for all specified optimizers,
\item a \code{character} (vector) of specified optimizer labels,
\item a \code{numeric} (vector) of optimizer ids, see the \code{$print()} output.
}}

\item{\code{only_comparable}}{Either \code{TRUE} to show only comparable results (i.e., results obtained
for the original optimization problem without any transformations),
or \code{FALSE} to include all optimization results.}

\item{\code{title}}{A \code{character}, the plot title.}

\item{\code{ylim}}{Passed on to \code{\link[ggplot2]{coord_cartesian}}.}

\item{\code{parameter_labels}}{A \code{character} of length \code{length(reference)} with labels for the
parameters.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{\link[ggplot2]{ggplot}} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Nop-trace"></a>}}
\if{latex}{\out{\hypertarget{method-Nop-trace}{}}}
\subsection{Method \code{trace()}}{
Capture trace of optimization with \code{stats::nlm()}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Nop$trace(
  initial = stats::rnorm(self$npar),
  iterations = 100,
  tolerance = 1e-06,
  which_element = c("value", "parameter", "gradient", "hessian", "seconds"),
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{initial}}{A \code{numeric} vector of length \code{npar}, the starting point for
the optimization.
By default, \code{initial = stats::rnorm(self$npar)}, i.e., random
initial values drawn from a standard normal distribution.}

\item{\code{iterations}}{A positive \code{integer}, the maximum number of iterations before
termination.
By default, \code{interations = 100}.}

\item{\code{tolerance}}{A \code{numeric}, the minimum allowable absolute change in the
function value before termination.
By default, \code{tolerance = 1e-6}.}

\item{\code{which_element}}{A \code{character} (vector) of elements to provide in the output, can be
one or more of:
\itemize{
\item \code{value} (the current function value)
\item \code{parameter} (the current value of each parameter)
\item \code{gradient} (the current gradient value)
\item \code{hessian} (the current Hessian value)
\item \code{seconds} (the number of seconds for the current iteration)
}}

\item{\code{...}}{Additional arguments passed on to \code{\link[stats]{nlm}}.
The arguments \code{iterlim} and \code{hessian} cannot be specified.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{data.frame} with iterations in rows, the columns depend on the
specification of \code{which_element}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Nop-best_value"></a>}}
\if{latex}{\out{\hypertarget{method-Nop-best_value}{}}}
\subsection{Method \code{best_value()}}{
Returns the best found \code{numeric} value of \code{f}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Nop$best_value(
  which_run = "all",
  which_optimizer = "all",
  only_comparable = TRUE,
  digits = getOption("ino_digits", default = 2),
  minimum = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{which_run}}{Selects saved results of optimization runs. Either:
\itemize{
\item \code{"all"} for all results,
\item \code{"last"} for the results of the last \code{$optimize()} call,
\item \code{"failed"}, the results from failed optimization runs,
\item a \code{character} (vector) of optimization labels,
\item an \code{integer} (vector) of optimization run ids.
}}

\item{\code{which_optimizer}}{Selects specified numerical optimizers. Either:
\itemize{
\item \code{"all"} for all specified optimizers,
\item a \code{character} (vector) of specified optimizer labels,
\item a \code{numeric} (vector) of optimizer ids, see the \code{$print()} output.
}}

\item{\code{only_comparable}}{Either \code{TRUE} to show only comparable results (i.e., results obtained
for the original optimization problem without any transformations),
or \code{FALSE} to include all optimization results.}

\item{\code{digits}}{An \code{integer}, the number of shown decimal places.}

\item{\code{minimum}}{TODO}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{numeric}, the best found \code{numeric} value of \code{f}.
The output has two attributes:
\itemize{
\item \code{run}, the run id that led to the best value,
\item \code{optimizer}, the optimizer that led to the best value.
In the case that multiple optimization runs led to the best value, the
first one of them is returned.
}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Nop-best_parameter"></a>}}
\if{latex}{\out{\hypertarget{method-Nop-best_parameter}{}}}
\subsection{Method \code{best_parameter()}}{
Returns the best found \code{numeric} parameter vector.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Nop$best_parameter(
  which_run = "all",
  which_optimizer = "all",
  only_comparable = TRUE,
  digits = getOption("ino_digits", default = 2),
  minimum = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{which_run}}{Selects saved results of optimization runs. Either:
\itemize{
\item \code{"all"} for all results,
\item \code{"last"} for the results of the last \code{$optimize()} call,
\item \code{"failed"}, the results from failed optimization runs,
\item a \code{character} (vector) of optimization labels,
\item an \code{integer} (vector) of optimization run ids.
}}

\item{\code{which_optimizer}}{Selects specified numerical optimizers. Either:
\itemize{
\item \code{"all"} for all specified optimizers,
\item a \code{character} (vector) of specified optimizer labels,
\item a \code{numeric} (vector) of optimizer ids, see the \code{$print()} output.
}}

\item{\code{only_comparable}}{Either \code{TRUE} to show only comparable results (i.e., results obtained
for the original optimization problem without any transformations),
or \code{FALSE} to include all optimization results.}

\item{\code{digits}}{An \code{integer}, the number of shown decimal places.}

\item{\code{minimum}}{TODO}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{numeric} of length \code{self$npar}.
The output has two attributes:
\itemize{
\item \code{run}, the run id that led to the best parameter vector,
\item \code{optimizer}, the optimizer that led to the best parameter vector.
Note that this parameter vector is not necessarily unique.
}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Nop-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Nop-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Nop$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
