% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nop.R
\name{Nop}
\alias{Nop}
\title{Nop Object (R6 Class)}
\description{
A \code{Nop} object defines a \emph{N}umerical \emph{O}ptimization \emph{P}roblem.
}
\section{Getting Started}{
\subsection{Step 1: Create a new \code{Nop} object:}{

Call \code{object <- Nop$new(f, npar, ...)} where
\itemize{
\item \code{f} is the function to be optimized over its first argument,
\item \code{npar} is the length of the first argument of \code{f},
\item and \code{...} are additional arguments for \code{f}.
}
}

\subsection{Step 2: Specify numerical optimizer:}{

Call \code{object$set_optimizer(<optimizer object>)}, where
\code{<optimizer object>} is an object of class \code{optimizer}, which can
be created with the \code{\link[optimizeR]{define_optimizer}} function from
the \{optimizeR\} package.
Two optimizer objects are already available:
\itemize{
\item \code{\link[optimizeR]{optimizer_nlm}}
\item \code{\link[optimizeR]{optimizer_optim}}
}
}

\subsection{Step 3: Test the configuration}{

Call \code{object$test()} to validate your configuration.
}
}

\section{Optimization}{
Call \code{object$optimize()} for optimization. See also the methods
\code{object$standardize()}, \code{object$reduce()}, and
\code{object$continue()} for initialization strategies.
}

\section{Analysis of the results}{
\itemize{
\item \code{object$results()} returns a \code{list} of all optimization results
\item \code{object$summary()} returns a \code{data.frame} with a summary of the
optimization results
\item \code{object$optima()} returns a \code{table} of identified optima
\item \code{object$plot()} visualizes an overview of optimization times
\item \code{object$best_parameter()} and \code{object$best_value()} return the
optimal parameter vector and optimal value over all optimization runs.
}
}

\examples{
Nop$new(f = f_ackley, npar = 2)$
  set_optimizer(optimizer_nlm())$
  optimize(runs = 100)$
  optima()

}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{f}}{The \code{function} to be optimized.}

\item{\code{npar}}{The length of the first argument of \code{f}.}

\item{\code{arguments}}{A \code{list} of specified additional arguments for \code{f}.}

\item{\code{true_parameter}}{The true optimum \code{numeric} parameter vector of length \code{npar} (if available).}

\item{\code{true_value}}{The true \code{numeric} optimum value of \code{f} (if available).}

\item{\code{show_minimum}}{A \code{logical}, set to \code{TRUE} (default) to
show best minimum in \code{$best_value} and \code{$best_parameter}.}

\item{\code{best_parameter}}{The best found \code{numeric} parameter vector of length \code{npar} (if available).}

\item{\code{best_value}}{The best found \code{numeric} value of \code{f} (if available).}

\item{\code{summary_columns}}{A \code{character}, the names of available statistics from the optimization.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Nop-new}{\code{Nop$new()}}
\item \href{#method-Nop-print}{\code{Nop$print()}}
\item \href{#method-Nop-set_argument}{\code{Nop$set_argument()}}
\item \href{#method-Nop-get_argument}{\code{Nop$get_argument()}}
\item \href{#method-Nop-remove_argument}{\code{Nop$remove_argument()}}
\item \href{#method-Nop-reset_argument}{\code{Nop$reset_argument()}}
\item \href{#method-Nop-evaluate}{\code{Nop$evaluate()}}
\item \href{#method-Nop-set_true_value}{\code{Nop$set_true_value()}}
\item \href{#method-Nop-set_true_parameter}{\code{Nop$set_true_parameter()}}
\item \href{#method-Nop-set_optimizer}{\code{Nop$set_optimizer()}}
\item \href{#method-Nop-remove_optimizer}{\code{Nop$remove_optimizer()}}
\item \href{#method-Nop-optimize}{\code{Nop$optimize()}}
\item \href{#method-Nop-test}{\code{Nop$test()}}
\item \href{#method-Nop-standardize}{\code{Nop$standardize()}}
\item \href{#method-Nop-reduce}{\code{Nop$reduce()}}
\item \href{#method-Nop-continue}{\code{Nop$continue()}}
\item \href{#method-Nop-summary}{\code{Nop$summary()}}
\item \href{#method-Nop-clear}{\code{Nop$clear()}}
\item \href{#method-Nop-optima}{\code{Nop$optima()}}
\item \href{#method-Nop-which_parameter}{\code{Nop$which_parameter()}}
\item \href{#method-Nop-plot}{\code{Nop$plot()}}
\item \href{#method-Nop-clone}{\code{Nop$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Nop-new"></a>}}
\if{latex}{\out{\hypertarget{method-Nop-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{Nop} (numerical optimization problem) object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Nop$new(f, npar, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{f}}{The \code{function} to be optimized.
It is optimized over its first argument, which should be a \code{numeric}
vector of length \code{npar}.}

\item{\code{npar}}{An \code{integer}, the length of the first argument of \code{f} (the
argument over which \code{f} is optimized).}

\item{\code{...}}{Optionally additional named arguments for \code{f}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{Nop} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Nop-print"></a>}}
\if{latex}{\out{\hypertarget{method-Nop-print}{}}}
\subsection{Method \code{print()}}{
Prints details of numerical optimization problem.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Nop$print(digits = getOption("ino_digits", default = 2), ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{digits}}{An \code{integer}, the number of decimal places.
The default is \code{2}.}

\item{\code{...}}{Currently not used.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Nop-set_argument"></a>}}
\if{latex}{\out{\hypertarget{method-Nop-set_argument}{}}}
\subsection{Method \code{set_argument()}}{
Set additional arguments for \code{f}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Nop$set_argument(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Optionally additional arguments for \code{f}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Nop-get_argument"></a>}}
\if{latex}{\out{\hypertarget{method-Nop-get_argument}{}}}
\subsection{Method \code{get_argument()}}{
Get value of an argument for \code{f}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Nop$get_argument(argument_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{argument_name}}{A \code{character}, the argument to extract.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Nop-remove_argument"></a>}}
\if{latex}{\out{\hypertarget{method-Nop-remove_argument}{}}}
\subsection{Method \code{remove_argument()}}{
Remove additional arguments for \code{f}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Nop$remove_argument(argument_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{argument_name}}{A \code{character} (vector), the argument(s) to remove.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Nop-reset_argument"></a>}}
\if{latex}{\out{\hypertarget{method-Nop-reset_argument}{}}}
\subsection{Method \code{reset_argument()}}{
Reset additional arguments for \code{f} after transformation with
\code{$standardize()} or \code{$reduce()}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Nop$reset_argument(
  argument_name,
  verbose = getOption("ino_verbose", default = TRUE)
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{argument_name}}{A \code{character} (vector), the argument(s) to reset.}

\item{\code{verbose}}{A \code{logical}, which indicates whether progress/details should be printed.
Set to \code{TRUE} (\code{FALSE}) to print (hide) such messages.
The default is \code{TRUE}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Nop-evaluate"></a>}}
\if{latex}{\out{\hypertarget{method-Nop-evaluate}{}}}
\subsection{Method \code{evaluate()}}{
Evaluate the function.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Nop$evaluate(at, time_limit = NULL, hide_warnings = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{at}}{A \code{numeric} vector of length \code{npar}.}

\item{\code{time_limit}}{An \code{integer}, the time limit in seconds for computations.
No time limit if \code{time_limit = NULL}.}

\item{\code{hide_warnings}}{A \code{logical}. Set to \code{TRUE} to hide any warnings.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Either:
\itemize{
\item a \code{numeric} value, the function value at \code{at},
\item \code{"time limit reached"} if the time limit was reached,
\item the error message if the evaluation failed.
}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Nop-set_true_value"></a>}}
\if{latex}{\out{\hypertarget{method-Nop-set_true_value}{}}}
\subsection{Method \code{set_true_value()}}{
Optionally set the true optimum value.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Nop$set_true_value(true_value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{true_value}}{A single \code{numeric}, the value of \code{f} at its optimum.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Nop-set_true_parameter"></a>}}
\if{latex}{\out{\hypertarget{method-Nop-set_true_parameter}{}}}
\subsection{Method \code{set_true_parameter()}}{
Optionally set the true optimum parameter vector.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Nop$set_true_parameter(true_parameter, set_true_value = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{true_parameter}}{A \code{numeric} vector of length \code{npar}, the point where \code{f}
obtains its optimum.}

\item{\code{set_true_value}}{Set to \code{TRUE} to use \code{true_parameter} to compute the true
optimum value of \code{f}.
Set to \code{FALSE} (default) if not.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Nop-set_optimizer"></a>}}
\if{latex}{\out{\hypertarget{method-Nop-set_optimizer}{}}}
\subsection{Method \code{set_optimizer()}}{
Set a numerical optimizer.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Nop$set_optimizer(optimizer, label = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{optimizer}}{An object of class \code{optimizer}, which can be created with the
\code{\link[optimizeR]{define_optimizer}} function from the \{optimizeR\}
package.}

\item{\code{label}}{A \code{character}, a unique label for the optimizer.
By default \code{label = NULL}, in which case the default label saved
inside \code{optimizer} is used.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Nop-remove_optimizer"></a>}}
\if{latex}{\out{\hypertarget{method-Nop-remove_optimizer}{}}}
\subsection{Method \code{remove_optimizer()}}{
Remove numerical optimizer.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Nop$remove_optimizer(which_optimizer = "all")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{which_optimizer}}{Select specified numerical optimizer. Either:
\itemize{
\item \code{"all"} for all specified optimizer (default)
\item a \code{character} (vector) of specified optimizer labels
\item a \code{numeric} (vector) of optimizer IDs (see the output of \code{$print()})
}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Nop-optimize"></a>}}
\if{latex}{\out{\hypertarget{method-Nop-optimize}{}}}
\subsection{Method \code{optimize()}}{
Optimize the function.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Nop$optimize(
  initial = "random",
  runs = 1,
  which_optimizer = "all",
  seed = NULL,
  return_results = FALSE,
  save_results = TRUE,
  label = "unlabeled",
  ncores = getOption("ino_ncores", default = 1),
  verbose = getOption("ino_verbose", default = TRUE),
  simplify = TRUE,
  reset_arguments_afterwards = TRUE,
  time_limit = NULL,
  hide_warnings = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{initial}}{Specify the initial point where the optimizer should start. Either:
\itemize{
\item the \code{character} \code{"random"} (the default) for random initial
values drawn from a standard normal distribution,
\item a \code{numeric} vector of length \code{npar}, the starting point for
optimization,
\item a \code{list} of such vectors (in this case, \code{runs} is set to the
length of the \code{list}),
\item or a \code{function} that returns a \code{numeric} vector of length
\code{npar} (the function can but not has to have a single integer
argument that specifies the optimization run).
}}

\item{\code{runs}}{An \code{integer}, the number of optimization runs.
By default, \code{runs = 1}.}

\item{\code{which_optimizer}}{Select specified numerical optimizer. Either:
\itemize{
\item \code{"all"} for all specified optimizer (default)
\item a \code{character} (vector) of specified optimizer labels
\item a \code{numeric} (vector) of optimizer IDs (see the output of \code{$print()})
}}

\item{\code{seed}}{Set a seed for reproducibility.
No seed by default.}

\item{\code{return_results}}{A \code{logical}, which indicates whether the results should be returned as
a \code{list}.
By default, \code{return_results = FALSE}.}

\item{\code{save_results}}{A \code{logical}, which indicates whether the results should be saved
inside the \code{Nop} object.
By default, \code{save_results = TRUE}.}

\item{\code{label}}{Only relevant if \code{save_results = TRUE}.
In this case, optionally a \code{character} for a custom label of the
optimization.
By default, \code{label = "unlabeled"}.
Labels can be useful to distinguish optimization runs later.}

\item{\code{ncores}}{An \code{integer}, the number of CPU cores for parallel computation.
The default is \code{1}.
You can use \code{parallel::detectCores()} to detect the number of available
CPU cores.}

\item{\code{verbose}}{A \code{logical}, which indicates whether progress/details should be printed.
Set to \code{TRUE} (\code{FALSE}) to print (hide) such messages.
The default is \code{TRUE}.}

\item{\code{simplify}}{Only relevant if \code{return_results = TRUE} and \code{runs = 1} and/or
only one optimizer is specified.
In this case, if \code{simplify = TRUE} (default), the nested list output
is flattened.}

\item{\code{reset_arguments_afterwards}}{A \code{logical}. Set to \code{TRUE} to reset all transformed
arguments after the optimization runs.
By default, \code{reset_arguments_afterwards = FALSE}.
Alternatively, reset arguments using \code{$reset_argument()}.}

\item{\code{time_limit}}{An \code{integer}, the time limit in seconds for computations.
No time limit if \code{time_limit = NULL}.}

\item{\code{hide_warnings}}{A \code{logical}. Set to \code{TRUE} to hide any warnings.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The return value depends on the value of \code{return_results}:
\itemize{
\item if \code{return_results = FALSE} (default), invisibly the \code{Nop}
object,
\item if \code{return_results = TRUE}, the output of
\code{\link[optimizeR]{apply_optimizer}}.
}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Nop-test"></a>}}
\if{latex}{\out{\hypertarget{method-Nop-test}{}}}
\subsection{Method \code{test()}}{
Test the configuration of a \code{Nop} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Nop$test(
  at = rnorm(self$npar),
  which_optimizer = "all",
  time_limit = 10,
  verbose = getOption("ino_verbose", default = TRUE),
  digits = getOption("ino_digits", default = 2)
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{at}}{A \code{numeric} of length \code{npar}, the point at which the
function \code{f} and the specified optimizer are tested.
Per default, \code{at = rnorm(self$npar)}, i.e., random values.}

\item{\code{which_optimizer}}{Select specified numerical optimizer. Either:
\itemize{
\item \code{"all"} for all specified optimizer (default)
\item a \code{character} (vector) of specified optimizer labels
\item a \code{numeric} (vector) of optimizer IDs (see the output of \code{$print()})
}}

\item{\code{time_limit}}{An \code{integer}, the time limit in seconds for computations.
No time limit if \code{time_limit = NULL}.}

\item{\code{verbose}}{A \code{logical}, which indicates whether progress/details should be printed.
Set to \code{TRUE} (\code{FALSE}) to print (hide) such messages.
The default is \code{TRUE}.}

\item{\code{digits}}{An \code{integer}, the number of decimal places.
The default is \code{2}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Invisibly \code{TRUE} if the tests are successful.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Nop-standardize"></a>}}
\if{latex}{\out{\hypertarget{method-Nop-standardize}{}}}
\subsection{Method \code{standardize()}}{
Standardize the optimization problem.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Nop$standardize(
  argument_name,
  by_column = TRUE,
  center = TRUE,
  scale = TRUE,
  ignore = integer(),
  verbose = getOption("ino_verbose", default = TRUE)
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{argument_name}}{A \code{character}, the name of the argument of \code{f} to be
standardized. The argument must be \code{numeric}.}

\item{\code{by_column}}{Only relevant if the argument \code{argument_name} is a \code{matrix} or
a \code{data.frame}.
In that case, either \code{TRUE} to standardize column-wise (default) or
\code{FALSE} to standardize row-wise.}

\item{\code{center}}{Passed to \code{\link[base]{scale}}.
Default is \code{TRUE}.}

\item{\code{scale}}{Passed to \code{\link[base]{scale}}.
Default is \code{TRUE}.}

\item{\code{ignore}}{Only relevant if the argument \code{argument_name} is a \code{matrix} or
a \code{data.frame}.
In that case, a \code{integer} (vector) of column indices (or row indices
if \code{by_column = FALSE}) to not standardize.}

\item{\code{verbose}}{A \code{logical}, which indicates whether progress/details should be printed.
Set to \code{TRUE} (\code{FALSE}) to print (hide) such messages.
The default is \code{TRUE}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Nop-reduce"></a>}}
\if{latex}{\out{\hypertarget{method-Nop-reduce}{}}}
\subsection{Method \code{reduce()}}{
Reduce the optimization problem.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Nop$reduce(
  argument_name,
  by_row = TRUE,
  how = "random",
  proportion = 0.5,
  centers = 2,
  ignore = integer(),
  seed = NULL,
  verbose = getOption("ino_verbose", default = TRUE)
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{argument_name}}{A \code{character}, the name of the argument of \code{f} to be reduced.}

\item{\code{by_row}}{Only relevant if the argument \code{argument_name} is a \code{matrix} or
a \code{data.frame}.
In that case, either \code{TRUE} to reduce row-wise (default) or
\code{FALSE} to reduce column-wise.}

\item{\code{how}}{A \code{character}, specifying how to reduce. Can be one of:
\itemize{
\item \code{"random"} (default), reduce at random
\item \code{"first"}, reduce to the first elements
\item \code{"last"}, reduce to the last elements
\item \code{"similar"}, reduce to similar elements
\item \code{"dissimilar"}, reduce to dissimilar elements
Note that \code{"similar"} and \code{"dissimilar"} are based on k-means
clustering via \code{\link[stats]{kmeans}}. To apply these options,
the argument \code{argument_name} must be \code{numeric}.
}}

\item{\code{proportion}}{A \code{numeric} between \code{0} and \code{1}, specifying the
reduction proportion.
By default, \code{proportion = 0.5}.}

\item{\code{centers}}{Only relevant, if \code{how = "(dis)similar"}.
In that case, passed to \code{\link[stats]{kmeans}}.
By default, \code{centers = 2}.}

\item{\code{ignore}}{Only relevant, if \code{how = "(dis)similar"} and the argument
\code{argument_name} is a \code{matrix} or a \code{data.frame}.
In that case a \code{integer} (vector) of row indices (or column indices
if \code{by_row = FALSE}) to ignore for clustering.}

\item{\code{seed}}{Set a seed for reproducibility.
No seed by default.}

\item{\code{verbose}}{A \code{logical}, which indicates whether progress/details should be printed.
Set to \code{TRUE} (\code{FALSE}) to print (hide) such messages.
The default is \code{TRUE}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Nop-continue"></a>}}
\if{latex}{\out{\hypertarget{method-Nop-continue}{}}}
\subsection{Method \code{continue()}}{
Continue last optimization runs, e.g., with a transformed parameter.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Nop$continue(
  save_results = TRUE,
  return_results = FALSE,
  time_limit = NULL,
  hide_warnings = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{save_results}}{A \code{logical}, which indicates whether the results should be saved
inside the \code{Nop} object.
By default, \code{save_results = TRUE}.}

\item{\code{return_results}}{A \code{logical}, which indicates whether the results should be returned as
a \code{list}.
By default, \code{return_results = FALSE}.}

\item{\code{time_limit}}{An \code{integer}, the time limit in seconds for computations.
No time limit if \code{time_limit = NULL}.}

\item{\code{hide_warnings}}{A \code{logical}. Set to \code{TRUE} to hide any warnings.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Nop-summary"></a>}}
\if{latex}{\out{\hypertarget{method-Nop-summary}{}}}
\subsection{Method \code{summary()}}{
Overview of the optimization runs.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Nop$summary(
  columns = c("value", "parameter", "seconds", "optimizer", "label"),
  which_runs = "all",
  which_optimizer = "all",
  digits = getOption("ino_digits", default = 2),
  only_comparable = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{columns}}{A \code{character} (vector), the names of columns to include in the
summary output.
The following columns are included by default:
\enumerate{
\item \code{"value"}, the value of the estimated function optimum
\item \code{"parameter"}, the parameter vector at which the optimum is
obtained
\item \code{"seconds"}, the optimization time in seconds
\item \code{"optimizer"}, the optimizer name
\item \code{"label"}, the optimization label
See \code{$summary_columns()} for an overview of the available column
names.
Set \code{columns = "all"} to include all of them in the output.
}}

\item{\code{which_runs}}{Select recorded results of optimization runs. Either:
\itemize{
\item \code{"all"} for all records (default)
\item \code{"last"} the records from the last optimization
\item a \code{character} (vector) of labels specified in \code{$optimize()}.
}}

\item{\code{which_optimizer}}{Select specified numerical optimizer. Either:
\itemize{
\item \code{"all"} for all specified optimizer (default)
\item a \code{character} (vector) of specified optimizer labels
\item a \code{numeric} (vector) of optimizer IDs (see the output of \code{$print()})
}}

\item{\code{digits}}{An \code{integer}, the number of decimal places.
The default is \code{2}.}

\item{\code{only_comparable}}{Either \code{TRUE} to show only comparable results (i.e., results obtained
for the original optimization problem), or \code{FALSE} to
include all optimization results.}

\item{\code{...}}{Optionally named expressions of variables from summary columns as
\code{character}.
See \code{$summary_columns()} for an overview of the available columns.
Also, \code{"true_value"} and \code{"true_parameter"} are available
(if specified).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{data.frame}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Nop-clear"></a>}}
\if{latex}{\out{\hypertarget{method-Nop-clear}{}}}
\subsection{Method \code{clear()}}{
Delete optimization results.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Nop$clear(which_runs = "all")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{which_runs}}{Select recorded results of optimization runs. Either:
\itemize{
\item \code{"all"} for all records (default)
\item \code{"last"} the records from the last optimization
\item a \code{character} (vector) of labels specified in \code{$optimize()}.
}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Nop-optima"></a>}}
\if{latex}{\out{\hypertarget{method-Nop-optima}{}}}
\subsection{Method \code{optima()}}{
Overview of the identified optima.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Nop$optima(
  digits = getOption("ino_digits", default = 2),
  sort_by = "frequency",
  which_runs = "all",
  which_optimizer = "all",
  only_comparable = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{digits}}{An \code{integer}, the number of decimal places.
The default is \code{2}.}

\item{\code{sort_by}}{Either \code{"frequency"} (default) to sort rows by frequency or
\code{"value"} to sort rows by value.}

\item{\code{which_runs}}{Select recorded results of optimization runs. Either:
\itemize{
\item \code{"all"} for all records (default)
\item \code{"last"} the records from the last optimization
\item a \code{character} (vector) of labels specified in \code{$optimize()}.
}}

\item{\code{which_optimizer}}{Select specified numerical optimizer. Either:
\itemize{
\item \code{"all"} for all specified optimizer (default)
\item a \code{character} (vector) of specified optimizer labels
\item a \code{numeric} (vector) of optimizer IDs (see the output of \code{$print()})
}}

\item{\code{only_comparable}}{Either \code{TRUE} to show only comparable results (i.e., results obtained
for the original optimization problem), or \code{FALSE} to
include all optimization results.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Nop-which_parameter"></a>}}
\if{latex}{\out{\hypertarget{method-Nop-which_parameter}{}}}
\subsection{Method \code{which_parameter()}}{
From all saved optimization results, extract the parameter vector that
led to a function value closest to \code{value}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Nop$which_parameter(
  value,
  which_runs = "all",
  which_optimizer = "all",
  only_comparable = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{A single \code{numeric}.}

\item{\code{which_runs}}{Select recorded results of optimization runs. Either:
\itemize{
\item \code{"all"} for all records (default)
\item \code{"last"} the records from the last optimization
\item a \code{character} (vector) of labels specified in \code{$optimize()}.
}}

\item{\code{which_optimizer}}{Select specified numerical optimizer. Either:
\itemize{
\item \code{"all"} for all specified optimizer (default)
\item a \code{character} (vector) of specified optimizer labels
\item a \code{numeric} (vector) of optimizer IDs (see the output of \code{$print()})
}}

\item{\code{only_comparable}}{Either \code{TRUE} to show only comparable results (i.e., results obtained
for the original optimization problem), or \code{FALSE} to
include all optimization results.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Nop-plot"></a>}}
\if{latex}{\out{\hypertarget{method-Nop-plot}{}}}
\subsection{Method \code{plot()}}{
Visualization of optimization time.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Nop$plot(by = NULL, relative = TRUE, log = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{by}}{Either:
\itemize{
\item \code{"label"} to group by optimization label
\item \code{"optimizer"} to group by optimizer
\item \code{NULL} to not group (default)
}}

\item{\code{relative}}{Either \code{TRUE} (default) to plot relative time differences with
respect to the median of the top boxplot, or \code{FALSE} to plot
absolute computation time.}

\item{\code{log}}{Either \code{TRUE} to plot a log10-x-axis, or \code{FALSE} (default)
if not.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Nop-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Nop-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Nop$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
