% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nop.R
\name{Nop}
\alias{Nop}
\title{Nop Object (R6 Class)}
\description{
A \code{Nop} object defines a numerical optimization problem.
}
\section{Initialize}{
TODO
\subsection{Step 1}{

TODO
}

\subsection{Step 2}{

TODO
}
}

\section{Optimize}{
TODO
}

\section{Evaluate}{
TODO
}

\examples{
ackley <- Nop$new(f = f_ackley, npar = 2)

}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{f}}{The \code{function} to be optimized.}

\item{\code{npar}}{The length of the first argument of \code{$f}.}

\item{\code{true_par}}{The true optimum parameter vector of length \code{$npar} (if available).}

\item{\code{true_val}}{The true optimum value of \code{$f} (if available).}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Nop-new}{\code{Nop$new()}}
\item \href{#method-Nop-print}{\code{Nop$print()}}
\item \href{#method-Nop-evaluate}{\code{Nop$evaluate()}}
\item \href{#method-Nop-optimize}{\code{Nop$optimize()}}
\item \href{#method-Nop-set_true_par}{\code{Nop$set_true_par()}}
\item \href{#method-Nop-set_true_val}{\code{Nop$set_true_val()}}
\item \href{#method-Nop-set_optimizer}{\code{Nop$set_optimizer()}}
\item \href{#method-Nop-remove_optimizer}{\code{Nop$remove_optimizer()}}
\item \href{#method-Nop-test}{\code{Nop$test()}}
\item \href{#method-Nop-clone}{\code{Nop$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Nop-new"></a>}}
\if{latex}{\out{\hypertarget{method-Nop-new}{}}}
\subsection{Method \code{new()}}{
Create a new numerical optimization problem.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Nop$new(f, npar)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{f}}{The \code{function} to be optimized.
It is optimized over its first argument.
Use \code{$add_pars()} to add additional parameters for \code{f}.}

\item{\code{npar}}{An \code{integer}, the length of the first argument of \code{f} (the
argument over which \code{f} is optimized).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{Nop} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Nop-print"></a>}}
\if{latex}{\out{\hypertarget{method-Nop-print}{}}}
\subsection{Method \code{print()}}{
Print method for numerical optimization problem.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Nop$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Currently not used.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Nop-evaluate"></a>}}
\if{latex}{\out{\hypertarget{method-Nop-evaluate}{}}}
\subsection{Method \code{evaluate()}}{
Evaluate the function.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Nop$evaluate(at)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{at}}{TODO}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Nop-optimize"></a>}}
\if{latex}{\out{\hypertarget{method-Nop-optimize}{}}}
\subsection{Method \code{optimize()}}{
Optimize the function.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Nop$optimize(initial = rnorm(self$npar), runs = 1, which_optimizer = "all")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{initial}}{TODO}

\item{\code{runs}}{TODO}

\item{\code{which_optimizer}}{Can be:
\itemize{
\item \code{"all"}
}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Nop-set_true_par"></a>}}
\if{latex}{\out{\hypertarget{method-Nop-set_true_par}{}}}
\subsection{Method \code{set_true_par()}}{
Optionally set the true optimum parameter.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Nop$set_true_par(true_par = NULL, set_true_val = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{true_par}}{Either \code{NULL} (default) or the point where \code{f} obtains its
optimum (i.e., a numeric vector of length \code{npar}).}

\item{\code{set_true_val}}{Set to \code{TRUE} to use \code{true_par} to compute the true
optimum value of \code{f} and to \code{FALSE} (default) if not.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Nop-set_true_val"></a>}}
\if{latex}{\out{\hypertarget{method-Nop-set_true_val}{}}}
\subsection{Method \code{set_true_val()}}{
Optionally set the true optimum value.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Nop$set_true_val(true_val = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{true_val}}{Either \code{NULL} (default) or the value of \code{f} at its optimum.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Nop-set_optimizer"></a>}}
\if{latex}{\out{\hypertarget{method-Nop-set_optimizer}{}}}
\subsection{Method \code{set_optimizer()}}{
Set a numerical optimizer.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Nop$set_optimizer(optimizer, label = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{optimizer}}{An object of class \code{optimizer} which can be created with the
\code{\link[optimizeR]{set_optimizer}} function from the {optimizeR}
package.}

\item{\code{label}}{A \code{character}, a unique label for the optimizer.
By default \code{NULL}, in which case the default label from
\code{optimizer} is used.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Nop-remove_optimizer"></a>}}
\if{latex}{\out{\hypertarget{method-Nop-remove_optimizer}{}}}
\subsection{Method \code{remove_optimizer()}}{
Remove a numerical optimizer.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Nop$remove_optimizer(id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{An \code{integer} (vector), the id(s) of optimizer to remove.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Nop-test"></a>}}
\if{latex}{\out{\hypertarget{method-Nop-test}{}}}
\subsection{Method \code{test()}}{
Test the configuration of numerical optimization problem.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Nop$test(at = rnorm(self$npar), time_limit = 3)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{at}}{A \code{numeric} of length \code{self$npar}, the point at which
function (\code{$f}) and the specified optimizer are tested.
Per default, \code{at = rnorm(self$npar)}.}

\item{\code{time_limit}}{A \code{numeric}, the time limit in seconds for testing the function
call and the optimization. If no error occurred after \code{time_limit}
seconds, the test is considered to be successful.
By default, \code{time_limit = 3}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Nop-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Nop-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Nop$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
