% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nop.R
\name{Nop}
\alias{Nop}
\title{Nop Object (R6 Class)}
\description{
A \code{Nop} object defines a \emph{n}umerical \emph{o}ptimization \emph{p}roblem.
}
\section{Getting Started}{
\subsection{Step 1: Create a new \code{Nop} object.}{

TODO
}

\subsection{Step 2: Add additional arguments.}{

TODO
}

\subsection{Step 2: Specify numerical optimizer.}{

TODO
}
}

\section{Optimize}{
TODO
}

\section{Evaluate}{
TODO
}

\examples{
ackley <- Nop$new(f = f_ackley, npar = 2)$
  set_optimizer(optimizer_nlm())$
  optimize()

}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{f}}{The \code{function} to be optimized.}

\item{\code{npar}}{The length of the first argument of \code{f}.}

\item{\code{arguments}}{A \code{list} of specified additional arguments for \code{f}.}

\item{\code{true_parameter}}{The true optimum parameter vector of length \code{npar} (if available).}

\item{\code{true_value}}{The true optimum value of \code{f} (if available).}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Nop-new}{\code{Nop$new()}}
\item \href{#method-Nop-print}{\code{Nop$print()}}
\item \href{#method-Nop-set_argument}{\code{Nop$set_argument()}}
\item \href{#method-Nop-get_argument}{\code{Nop$get_argument()}}
\item \href{#method-Nop-remove_argument}{\code{Nop$remove_argument()}}
\item \href{#method-Nop-evaluate}{\code{Nop$evaluate()}}
\item \href{#method-Nop-set_true_value}{\code{Nop$set_true_value()}}
\item \href{#method-Nop-set_true_parameter}{\code{Nop$set_true_parameter()}}
\item \href{#method-Nop-set_optimizer}{\code{Nop$set_optimizer()}}
\item \href{#method-Nop-remove_optimizer}{\code{Nop$remove_optimizer()}}
\item \href{#method-Nop-optimize}{\code{Nop$optimize()}}
\item \href{#method-Nop-test}{\code{Nop$test()}}
\item \href{#method-Nop-standardize}{\code{Nop$standardize()}}
\item \href{#method-Nop-reduce}{\code{Nop$reduce()}}
\item \href{#method-Nop-optima}{\code{Nop$optima()}}
\item \href{#method-Nop-summary}{\code{Nop$summary()}}
\item \href{#method-Nop-plot}{\code{Nop$plot()}}
\item \href{#method-Nop-clone}{\code{Nop$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Nop-new"></a>}}
\if{latex}{\out{\hypertarget{method-Nop-new}{}}}
\subsection{Method \code{new()}}{
Create a new numerical optimization problem.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Nop$new(f, npar, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{f}}{The \code{function} to be optimized.
It is optimized over its first argument, which should be a \code{numeric}
vector of length \code{npar}.}

\item{\code{npar}}{An \code{integer}, the length of the first argument of \code{f} (the
argument over which \code{f} is optimized).}

\item{\code{...}}{Optionally additional arguments for \code{f}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{Nop} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Nop-print"></a>}}
\if{latex}{\out{\hypertarget{method-Nop-print}{}}}
\subsection{Method \code{print()}}{
Print method for numerical optimization problem.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Nop$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Currently not used.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Nop-set_argument"></a>}}
\if{latex}{\out{\hypertarget{method-Nop-set_argument}{}}}
\subsection{Method \code{set_argument()}}{
Set additional arguments for \code{f}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Nop$set_argument(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Optionally additional arguments for \code{f}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Nop-get_argument"></a>}}
\if{latex}{\out{\hypertarget{method-Nop-get_argument}{}}}
\subsection{Method \code{get_argument()}}{
Get value of additional argument for \code{f}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Nop$get_argument(argument_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{argument_name}}{A \code{character}, the argument to extract.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Nop-remove_argument"></a>}}
\if{latex}{\out{\hypertarget{method-Nop-remove_argument}{}}}
\subsection{Method \code{remove_argument()}}{
Remove additional arguments for \code{f}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Nop$remove_argument(argument_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{argument_name}}{A \code{character}, the argument to remove.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Nop-evaluate"></a>}}
\if{latex}{\out{\hypertarget{method-Nop-evaluate}{}}}
\subsection{Method \code{evaluate()}}{
Evaluate the function.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Nop$evaluate(at)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{at}}{A \code{numeric} vector of length \code{npar}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{numeric} value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Nop-set_true_value"></a>}}
\if{latex}{\out{\hypertarget{method-Nop-set_true_value}{}}}
\subsection{Method \code{set_true_value()}}{
Optionally set the true optimum value.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Nop$set_true_value(true_value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{true_value}}{A single \code{numeric}, the value of \code{f} at its optimum.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Nop-set_true_parameter"></a>}}
\if{latex}{\out{\hypertarget{method-Nop-set_true_parameter}{}}}
\subsection{Method \code{set_true_parameter()}}{
Optionally set the true optimum parameter vector.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Nop$set_true_parameter(true_parameter, set_true_value = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{true_parameter}}{A \code{numeric} vector of length \code{npar}, the point where \code{f}
obtains its optimum.}

\item{\code{set_true_value}}{Set to \code{TRUE} to use \code{true_parameter} to compute the true
optimum value of \code{f}.
Set to \code{FALSE} (default) if not.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Nop-set_optimizer"></a>}}
\if{latex}{\out{\hypertarget{method-Nop-set_optimizer}{}}}
\subsection{Method \code{set_optimizer()}}{
Set a numerical optimizer.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Nop$set_optimizer(optimizer, label = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{optimizer}}{An object of class \code{optimizer}, which can be created with the
\code{\link[optimizeR]{set_optimizer}} function from the \{optimizeR\}
package.}

\item{\code{label}}{A \code{character}, a unique label for the optimizer.
By default \code{label = NULL}, in which case the default label saved
inside \code{optimizer} is used.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Nop-remove_optimizer"></a>}}
\if{latex}{\out{\hypertarget{method-Nop-remove_optimizer}{}}}
\subsection{Method \code{remove_optimizer()}}{
Remove numerical optimizer.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Nop$remove_optimizer(which_optimizer = "all")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{which_optimizer}}{Select specified numerical optimizer. Either:
\itemize{
\item \code{"all"} for all specified optimizer (default)
\item a \code{character} vector of specified optimizer labels
\item a \code{numeric} vector of optimizer IDs (see the output of \code{$print()})
}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Nop-optimize"></a>}}
\if{latex}{\out{\hypertarget{method-Nop-optimize}{}}}
\subsection{Method \code{optimize()}}{
Optimize the function.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Nop$optimize(
  initial = "random",
  runs = 1,
  which_optimizer = "all",
  seed = NULL,
  save_results = TRUE,
  return_results = FALSE,
  label = "",
  ncores = getOption("ino_ncores"),
  verbose = getOption("ino_progress"),
  simplify = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{initial}}{Specify the initial point where the optimizer should start. Either:
\itemize{
\item the character \code{"random"} (the default) for random initial values
drawn from a standard normal distribution
\item a \code{numeric} vector of length \code{npar}, the starting point for
optimization
\item a \code{function} without any arguments that returns a \code{numeric}
vector of length \code{npar}
}}

\item{\code{runs}}{An \code{integer}, the number of optimization runs.
By default, \code{runs = 1}.}

\item{\code{which_optimizer}}{Select specified numerical optimizer. Either:
\itemize{
\item \code{"all"} for all specified optimizer (default)
\item a \code{character} vector of specified optimizer labels
\item a \code{numeric} vector of optimizer IDs (see the output of \code{$print()})
}}

\item{\code{seed}}{Set a seed. No seed by default.}

\item{\code{save_results}}{A \code{logical}, which indicates whether the results should be saved
inside the \code{Nop} object.
By default, \code{save_results = TRUE}.}

\item{\code{return_results}}{A \code{logical}, which indicates whether the results should be returned.
By default, \code{return_results = FALSE}.}

\item{\code{label}}{TODO}

\item{\code{ncores}}{An \code{integer}, the number of cores for parallel computation.
The default is \code{getOption("ino_ncores")}, which is set to \code{1}
when the package is loaded.}

\item{\code{verbose}}{A \code{logical}, which indicates whether progress should be printed.
Set to \code{TRUE} (\code{FALSE}) to print (hide) progress.
The default is \code{getOption("ino_progress")}, which is set to \code{TRUE}
when the package is loaded.}

\item{\code{simplify}}{TODO}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The return value depends on the value of \code{return_results}:
\itemize{
\item if \code{return_results = FALSE} (default), invisibly the \code{Nop} object
\item if \code{return_results = TRUE}, the output of \code{\link[optimizeR]{apply_optimizer}}.
}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Nop-test"></a>}}
\if{latex}{\out{\hypertarget{method-Nop-test}{}}}
\subsection{Method \code{test()}}{
Test the configuration of numerical optimization problem.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Nop$test(at = rnorm(self$npar), which_optimizer = "all", time_limit = 10)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{at}}{A \code{numeric} of length \code{npar}, the point at which the
function \code{f} and the specified optimizer are tested.
Per default, \code{at = rnorm(self$npar)}, i.e., random values.}

\item{\code{which_optimizer}}{Select specified numerical optimizer. Either:
\itemize{
\item \code{"all"} for all specified optimizer (default)
\item a \code{character} vector of specified optimizer labels
\item a \code{numeric} vector of optimizer IDs (see the output of \code{$print()})
}}

\item{\code{time_limit}}{A \code{numeric}, the time limit in seconds for testing the function
call and the optimization.
If no error occurred after \code{time_limit} seconds, the test is
considered to be successful.
By default, \code{time_limit = 10}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Nop-standardize"></a>}}
\if{latex}{\out{\hypertarget{method-Nop-standardize}{}}}
\subsection{Method \code{standardize()}}{
Standardize the optimization problem.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Nop$standardize(
  argument_name,
  by_column = TRUE,
  center = TRUE,
  scale = TRUE,
  ignore = integer()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{argument_name}}{A \code{character}, the name of the argument of \code{f} to be standardized.
The argument must be either a \code{data.frame} or a \code{matrix}.}

\item{\code{by_column}}{Either \code{TRUE} to standardize column-wise (default) or \code{FALSE}
to standardize row-wise.}

\item{\code{center}}{Passed to \code{\link[base]{scale}}. Default is \code{TRUE}.}

\item{\code{scale}}{Passed to \code{\link[base]{scale}}. Default is \code{TRUE}.}

\item{\code{ignore}}{A \code{integer} (vector) of column indices (or row indices if
\code{by_column = FALSE}) to not standardize.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Nop-reduce"></a>}}
\if{latex}{\out{\hypertarget{method-Nop-reduce}{}}}
\subsection{Method \code{reduce()}}{
Reduce the optimization problem.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Nop$reduce(
  argument_name,
  by_row = TRUE,
  how = "random",
  proportion = 0.5,
  centers = 2,
  ignore = integer(),
  seed = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{argument_name}}{A \code{character}, the name of the argument of \code{f} to be subsetted.
The argument must be either a \code{data.frame} or a \code{matrix}.}

\item{\code{by_row}}{Either \code{TRUE} to subset row-wise (default) or \code{FALSE}
to subset column-wise.}

\item{\code{how}}{A \code{character}, specifying how to select the subset. Can be one of:
\itemize{
\item \code{"random"} (default)
\item \code{"first"}
\item \code{"last"}
\item \code{"similar"}
\item \code{"unsimilar"}
}}

\item{\code{proportion}}{A \code{numeric} between \code{0} and \code{1}, specifying the proportion
of the subset.
By default, \code{proportion = 0.5}.}

\item{\code{centers}}{Only relevant if \code{how = "(un)similar"}.
In that case, passed to \code{\link[stats]{kmeans}}.
By default, \code{centers = 2}.}

\item{\code{ignore}}{Only relevant if \code{how = "(un)similar"}.
In that case a \code{integer} (vector) of row indices (or column indices
if \code{by_row = FALSE}) to ignore for clustering.}

\item{\code{seed}}{Set a seed. No seed by default.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Nop-optima"></a>}}
\if{latex}{\out{\hypertarget{method-Nop-optima}{}}}
\subsection{Method \code{optima()}}{
TODO
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Nop$optima(digits = 2)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{digits}}{TODO}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Nop-summary"></a>}}
\if{latex}{\out{\hypertarget{method-Nop-summary}{}}}
\subsection{Method \code{summary()}}{
TODO
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Nop$summary()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Nop-plot"></a>}}
\if{latex}{\out{\hypertarget{method-Nop-plot}{}}}
\subsection{Method \code{plot()}}{
TODO
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Nop$plot()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Nop-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Nop-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Nop$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
