% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nop.R
\name{Nop}
\alias{Nop}
\title{Nop Object}
\description{
A \code{Nop} object defines a numerical optimization problem.
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{npar}}{[\code{integer()}, read-only]\cr
The length of each target argument.}

\item{\code{verbose}}{[\code{logical(1)}]\cr
Print progress and details?}

\item{\code{digits}}{verbose [\code{integer(1)}]\cr
The number of decimal places of interest.}

\item{\code{fresh_label}}{[\code{character(1)}, read-only]\cr
An optimization label that has not been used yet.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Nop-new}{\code{Nop$new()}}
\item \href{#method-Nop-print}{\code{Nop$print()}}
\item \href{#method-Nop-evaluate}{\code{Nop$evaluate()}}
\item \href{#method-Nop-set_optimizer}{\code{Nop$set_optimizer()}}
\item \href{#method-Nop-initialize_fixed}{\code{Nop$initialize_fixed()}}
\item \href{#method-Nop-initialize_random}{\code{Nop$initialize_random()}}
\item \href{#method-Nop-initialize_custom}{\code{Nop$initialize_custom()}}
\item \href{#method-Nop-optimize}{\code{Nop$optimize()}}
\item \href{#method-Nop-clone}{\code{Nop$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Nop-new"></a>}}
\if{latex}{\out{\hypertarget{method-Nop-new}{}}}
\subsection{Method \code{new()}}{
Creates a new \code{Nop} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Nop$new(f, target = NULL, npar, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{f}}{[\code{function}]\cr
A \code{function} to be optimized (the so-called objective function).

It is expected that \code{f} has at least one \code{numeric} argument.

Further, it is expected that the return value of \code{f} is of the
structure \code{numeric(1)}, i.e. a single \code{numeric} value.}

\item{\code{target}}{[\code{character()}]\cr
The argument name(s) that get optimized (the so-called target arguments).

All target arguments must be \code{numeric}.

Can be \code{NULL} (default), then the first function argument is selected.}

\item{\code{npar}}{[\code{integer()}]\cr
The length of each target argument, i.e., the length(s) of the
\code{numeric} \code{vector} argument(s) specified by \code{target}.}

\item{\code{...}}{Optionally additional function arguments that are fixed during the
optimization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Nop-print"></a>}}
\if{latex}{\out{\hypertarget{method-Nop-print}{}}}
\subsection{Method \code{print()}}{
Prints details of the \code{Nop} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Nop$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Currently not used.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Nop-evaluate"></a>}}
\if{latex}{\out{\hypertarget{method-Nop-evaluate}{}}}
\subsection{Method \code{evaluate()}}{
Evaluates the objective function.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Nop$evaluate(at = stats::rnorm(sum(self$npar)))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{at}}{[\code{numeric()}]\cr
The values for the target argument(s), written in a single vector.

Must be of length \code{sum(self$npar)}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Nop-set_optimizer"></a>}}
\if{latex}{\out{\hypertarget{method-Nop-set_optimizer}{}}}
\subsection{Method \code{set_optimizer()}}{
Specifies a numerical optimizer.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Nop$set_optimizer(optimizer, optimizer_label = optimizer$label)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{optimizer}}{[\code{Optimizer}]\cr
An \code{Optimizer} object, which can be created via
\code{\link[optimizeR]{Optimizer}}.}

\item{\code{optimizer_label}}{[\code{character(1)}]\cr
A (unique) label for the optimizer.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Nop-initialize_fixed"></a>}}
\if{latex}{\out{\hypertarget{method-Nop-initialize_fixed}{}}}
\subsection{Method \code{initialize_fixed()}}{
Defines fixed initial values for the optimization.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Nop$initialize_fixed(at)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{at}}{[\code{integer(self$sum(npar))} | \code{list()}]\cr
The fixed initial parameter vector.

It can also be a \code{list} of such vectors.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Nop-initialize_random"></a>}}
\if{latex}{\out{\hypertarget{method-Nop-initialize_random}{}}}
\subsection{Method \code{initialize_random()}}{
Defines random initial values for the optimization.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Nop$initialize_random(
  sampler = function() stats::rnorm(sum(self$npar)),
  runs = 1L
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sampler}}{[\code{function}]\cr
A \code{function} without any arguments that returns a \code{numeric}
vector of length \code{sum(self$npar)}.}

\item{\code{runs}}{[\code{integer(1)}]\cr
The number of optimization runs.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Nop-initialize_custom"></a>}}
\if{latex}{\out{\hypertarget{method-Nop-initialize_custom}{}}}
\subsection{Method \code{initialize_custom()}}{
Defines custom initial values for the optimization.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Nop$initialize_custom(at, seconds = rep(0, length(at)), type = "custom")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{at}}{[\code{list()}]\cr
A \code{list} of initial parameter vectors.}

\item{\code{seconds}}{[\code{numeric(length(at))}]\cr
The number of seconds it took to obtain each initial value in \code{at},
which is added to the overall optimization time.}

\item{\code{type}}{[\code{character(1)}]\cr
The type of the initial values.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Nop-optimize"></a>}}
\if{latex}{\out{\hypertarget{method-Nop-optimize}{}}}
\subsection{Method \code{optimize()}}{
Optimizes the target function.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Nop$optimize(
  optimization_label = self$fresh_label,
  which_optimizer = "all",
  which_direction = "min",
  seconds = Inf,
  hide_warnings = TRUE,
  reset_initial_afterwards = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{optimization_label}}{[\code{character(1)}]\cr
A label for the optimization to distinguish optimization runs.}

\item{\code{reset_initial_afterwards}}{[\code{logical(1)}]\cr
Reset the initial values after the optimization?}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Supports:
\itemize{
\item Parallel computation of multiple optimization runs via \code{{future}}
\item Progress messages via \code{{progressr}}
}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Nop-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Nop-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Nop$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
