% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{nr_optima}
\alias{nr_optima}
\title{Overview of the number of optima found.}
\usage{
nr_optima(x, plot = FALSE, round = 2)
}
\arguments{
\item{x}{An object of class \code{ino}.}

\item{plot}{Set to \code{TRUE} for plotting a frequency plot of found optima.}

\item{round}{Sets the number of decimal places for the location of the optima.}
}
\value{
A data frame.
}
\description{
This function provides an overview of the number of optima found with respect to
the initialization strategies.
}
\examples{
# nr_optima()

}
\keyword{evaluation}
