% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization_time.R
\name{optimization_time}
\alias{optimization_time}
\title{Overview of optimization times.}
\usage{
optimization_time(x, plot_hist = FALSE, plot_freq = FALSE, round = 2)
}
\arguments{
\item{x}{An object of class \code{ino}.}

\item{plot_hist}{Set to \code{TRUE} for plotting a histogram of the optimization times.}

\item{plot_freq}{Set to \code{TRUE} for plotting a frequency plot of the optimization times.}

\item{round}{Sets the number of decimal places for the optimization time.}
}
\value{
A data frame.
}
\description{
This function provides an overview of the optimization times with respect to
the initialization strategies.
}
\examples{
# optimization_time()

}
\keyword{evaluation}
