% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategies.R
\name{optimize}
\alias{optimize}
\title{Optimization}
\usage{
optimize(x, init, ncores, verbose)
}
\arguments{
\item{x}{An object of class \code{ino}.}

\item{init}{A numeric vector of length \code{npar(x)}.}

\item{ncores}{The number of cores for parallel computation over parameters and optimizers.
The default is \code{getOption("ino_ncores")}, which is set to \code{1}
when the package is loaded.}

\item{verbose}{A boolean, which indicates whether progress should be printed.
Set to \code{TRUE} (\code{FALSE}) to print (hide) progress.
The default is \code{getOption("ino_progress")}, which is set to \code{TRUE}
when the package is loaded.}
}
\value{
A list, each element contains
\itemize{
\item the set number \code{i} corresponding to \code{grid_ino(x)}
\item and the list output of \code{\link[optimizeR]{optimizeR}}.
}
}
\description{
This helper function performs numerical optimization based on an \code{ino}
object and initial values.
}
\keyword{internal}
