% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizer.R
\name{optimizer_ao}
\alias{optimizer_ao}
\title{Specify ao optimizer}
\usage{
optimizer_ao(
  ...,
  output_ignore = character(),
  validate = TRUE,
  validation_settings = list()
)
}
\arguments{
\item{...}{Additional arguments to be passed to the optimizer. Without
specifications, the default values of the optimizer are used.}

\item{output_ignore}{A \code{character} vector of element names in the output of \code{optimizer}
that are not saved. The elements \code{value} and \code{parameter} are
added automatically to \code{output_ignore}, because they are saved
separately, see the output documentation of \code{\link[optimizeR]{apply_optimizer}}.}

\item{validate}{A \code{logical}, set to \code{TRUE} (\code{FALSE}) to (not) validate the
\code{optimizer} object.
By default, \code{validate = TRUE}.}

\item{validation_settings}{Ignored if \code{valdiate = FALSE}.
Otherwise, a \code{list} of validation settings:
\describe{
\item{objective_test}{A \code{function}, the test function to be optimized.
By default, it is the
\href{https://en.wikipedia.org/wiki/Ackley_function}{Ackley function}.}
\item{objective_add}{A \code{list} of additional arguments to
\code{objective_test} (if any).
By default, \code{objective_add = list()}, because the default function
for \code{objective_test} does not have additional arguments.}
\item{initial}{A \code{numeric} vector, the initial values for the
optimization of \code{objective_test}.
By default, \code{initial = round(stats::rnorm(2), 2)}.}
\item{check_seconds}{An \code{integer}, the maximum number of seconds
before the test is aborted.
The test call is considered to be successful if no error occurred
within \code{check_seconds} seconds.
By default, \code{check_seconds = 10}.}
}}
}
\value{
An object of class \code{optimizer}.
}
\description{
This function is a wrapper for \code{\link[optimizeR]{define_optimizer}} with
the \code{\link[ao]{ao}} optimizer.
}
