% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluations.R
\name{overview_optima}
\alias{overview_optima}
\title{Optima overview}
\usage{
overview_optima(x, digits = 2)
}
\arguments{
\item{x}{An object of class \code{ino}.}

\item{digits}{The number of digits of the optima values.}
}
\value{
A data frame.
}
\description{
This function provides an overview of the identified optima.
}
\keyword{evaluation}
