% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluations.R
\name{plot.ino}
\alias{plot.ino}
\title{Visualization of optimization time}
\usage{
\method{plot}{ino}(x, by = NULL, time_unit = "secs", ...)
}
\arguments{
\item{x}{An object of class \code{ino}.}

\item{by}{A character vector of variables to group by. Can be \code{NULL} (default).}

\item{time_unit}{The time unit, see \code{\link{difftime}}.}

\item{...}{Ignored.}
}
\value{
A \code{ggplot} object.
}
\description{
This function plots boxplots of optimization times in an \code{ino} object.
}
\keyword{evaluation}
