% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sets.R
\docType{data}
\name{probit_data}
\alias{probit_data}
\title{Simulated choice data from a probit model}
\format{
The format of the data sets is documented in \code{\link{sim_mnp}}.
}
\usage{
data(probit_data)
}
\description{
This object is a \code{list} which includes 100 data sets of simulated
choices and corresponding choice characteristics from a probit model.
}
\details{
The data set was simulated via the following lines:
\preformatted{
  set.seed(1)
  probit_data <- list()
  for(i in 1:100){
    b <- c(1,rnorm(2, sd = 3))
    Sigma <- RprobitB::rwishart(3,diag(3))$W
    name <- paste0("data",i)
    probit_data[[name]] <- ino:::sim_mnp(
       N = 100, b = b, Sigma = Sigma, seed = i
       )
  }
}
}
\keyword{dataset}
