% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ino_objects.R
\docType{data}
\name{probit_ino}
\alias{probit_ino}
\title{Probit ino object}
\format{
Missing
}
\usage{
data(probit_ino)
}
\description{
Missing
}
\details{
Missing
\preformatted{
  NA
}
}
\keyword{dataset}
