% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demos.R
\docType{data}
\name{probit_ino}
\alias{probit_ino}
\title{Example application of to probit likelihood}
\format{
A \code{\link{Nop}} object.
}
\usage{
data(probit_ino)
}
\description{
See the vignette about the probit likelihood for details:
\url{https://loelschlaeger.de/ino/articles/example_probit.html}
}
\keyword{demo}
