% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategies.R
\name{random_initialization}
\alias{random_initialization}
\title{Random initialization}
\usage{
random_initialization(x, runs = 1, sampler = stats::rnorm)
}
\arguments{
\item{x}{An object of class \code{ino}.}

\item{runs}{The number of random initialization runs.}

\item{sampler}{If unspecified, draws independent initial values from a standard normal
distribution. Else can be a function that returns a (random) numeric vector
of length \code{npar}.}
}
\value{
The updated input \code{x}.
}
\description{
This function is an implementation of the random initialization
strategy. See the vignette on initialization strategies for more details.
}
\details{
Missing.
}
\examples{
#random_initialization()

}
\keyword{strategy}
