% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategies.R
\name{random_initialization}
\alias{random_initialization}
\title{Random initialization}
\usage{
random_initialization(x, runs = 1, sampler = NULL)
}
\arguments{
\item{x}{An object of class \code{ino}.}

\item{runs}{The number of random initialization runs.}

\item{sampler}{If unspecified, draws independent initial values from a standard normal
distribution. Else can be a function that returns a (random) numeric vector
of length \code{npar}.}
}
\value{
The \code{ino} object.
}
\description{
This function implemented the random initialization strategy.
}
\details{
See the vignette ... for more details.
}
\examples{
#random_initialization()

}
\keyword{strategy}
