% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategies.R
\name{random_initialization}
\alias{random_initialization}
\title{Random initialization}
\usage{
random_initialization(x, runs = 1, sampler = stats::rnorm, ...)
}
\arguments{
\item{x}{An object of class \code{ino}.}

\item{runs}{The number of random initialization runs, the default is \code{1}.}

\item{sampler}{The sampler for random initial values. Can be any function that
\itemize{
\item has as first argument an integer, say \code{npar},
\item and returns a numeric vector of length \code{npar}.
}
Per default, \code{sampler = stats::rnorm}, i.e. independent draws from a
standard normal distribution as initial value.}

\item{...}{Additional argument to \code{sampler} (optional).}
}
\value{
The updated input \code{x}.
}
\description{
This function is an implementation of the random initialization
strategy.
}
\details{
See the vignette on initialization strategies for more details.
}
\examples{
#random_initialization()

}
\keyword{strategy}
