% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{remove_index}
\alias{remove_index}
\title{Helper function for ...}
\usage{
remove_index(x, index, replace = FALSE)
}
\arguments{
\item{x}{TODO}

\item{index}{TODO}

\item{replace}{Either \code{TRUE} ...}
}
\value{
TODO
}
\description{
This function ... TODO
}
\keyword{internal}
