% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runs.R
\name{Runs}
\alias{Runs}
\title{Runs Object (R6 Class)}
\value{
A \code{Runs} object, which is an R6 class that stores results of numerical
optimization runs, see the details.
}
\description{
A \code{Runs} object stores results of numerical optimization runs. It is
typically contained inside a \code{\link{Nop}} object.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Runs-new}{\code{Runs$new()}}
\item \href{#method-Runs-print}{\code{Runs$print()}}
\item \href{#method-Runs-clone}{\code{Runs$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Runs-new"></a>}}
\if{latex}{\out{\hypertarget{method-Runs-new}{}}}
\subsection{Method \code{new()}}{
Creates a new \code{Runs} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Runs$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A new \code{Runs} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Runs-print"></a>}}
\if{latex}{\out{\hypertarget{method-Runs-print}{}}}
\subsection{Method \code{print()}}{
Prints details of the stored numerical optimization results.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Runs$print(digits = getOption("ino_digits", default = 2), ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{digits}}{An \code{integer}, the number of shown decimal places.
The default is \code{2}.}

\item{\code{...}}{Currently not used.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Invisibly the \code{Results} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Runs-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Runs-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Runs$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
