% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategies.R
\name{save_result}
\alias{save_result}
\title{Save results}
\usage{
save_result(x, result, strategy, init)
}
\arguments{
\item{x}{An object of class \code{ino}.}

\item{result}{The output of \code{\link{optimize}}.}

\item{strategy}{The name of the initialization strategy.}

\item{A}{numeric vector of initial parameters of length \code{npar(x)}.}
}
\value{
The updated input \code{x}.
}
\description{
This helper function saves the results of optimization runs into the
submitted \code{ino} object.
}
\details{
The results are saved at \code{x$runs}, which is an object of class
\code{runs}. See \code{\link{new_runs}} for its documentation.
}
\keyword{internal}
