% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_f.R
\name{set_f}
\alias{set_f}
\title{Set a function for the numerical optimization.}
\usage{
set_f(f, npar)
}
\arguments{
\item{f}{An object of class \code{function}.}

\item{npar}{The number of parameters in \code{f}.}
}
\value{
An object of class \code{ino_f}.
}
\description{
Set a function for the numerical optimization.
}
\examples{
f <- function(x) x
set_f(f, 1)
}
