% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{set_optimizer}
\alias{set_optimizer}
\title{Specify a numerical optimizer for an \code{ino} object}
\usage{
set_optimizer(opt, f, p, z, ..., crit = c(z))
}
\arguments{
\item{opt}{An object of class \code{function}, a numerical optimizer, see the details.}

\item{f}{The name of the function input of \code{opt}.}

\item{p}{The name of the starting parameter values input of \code{opt}.}

\item{z}{The name of the optimal parameter vector in the output list of \code{opt}.}

\item{...}{Additional arguments to be passed to the optimizer \code{opt}. Without
specifications, the default values of \code{opt} are used.}

\item{crit}{The names of the list elements in the output of \code{opt} to be saved after
the optimization. Must contain \code{z}.}
}
\value{
An object of class \code{optimizer}, which can be passed to
\code{\link{setup_ino}}.
}
\description{
Use this function to specify a numerical optimizer for the optimization
problem.
}
\details{
Numerical optimization functions specified for the \code{opt} argument must
fulfill the following requirements:
\itemize{
\item function input name f
\item starting parameter values input name p
\item additional parameters to f via ...
\item output is a named list
\item name of optimal parameter vector in the output list is z
}
}
\examples{
set_optimizer(
  opt = pracma::nelder_mead,
  f = "fn",
  p = "x0",
  z = "xmin",
  tol = 1e-6,
  crit = c("xmin", "fcount")
)

}
\seealso{
\code{\link[=set_optimizer_nlm]{set_optimizer_nlm()}}, a wrapper for the \code{\link[stats]{nlm}} optimizer.
}
\keyword{specification}
