% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_optimizer.R
\name{set_optimizer}
\alias{set_optimizer}
\title{Title}
\usage{
set_optimizer(x, optimizer, ...)
}
\arguments{
\item{x}{An object of class \code{ino}.}

\item{optimizer}{An object of class \code{function}.}

\item{...}{Additional arguments to be passed to \code{optimizer}.}
}
\description{
Title
}
\examples{
# set_optimizer()
}
