% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{simplify_results}
\alias{simplify_results}
\title{Simplify optimization results}
\usage{
simplify_results(results, simplify)
}
\arguments{
\item{results}{A nested \code{list} of optimization results.
Each element corresponds to one optimization run.
It is either \code{list()} if the run has been removed or a \code{list}
of results for each optimizer.
The results for each optimizer is a \code{list}, the output of
\code{\link[optimizeR]{apply_optimizer}}.}

\item{simplify}{See documentation of method \code{$results()} from \code{Nop} object.}
}
\value{
A \code{list}.
}
\description{
This helper function simplifies optimization results (if possible).
}
\examples{
\dontrun{
simplify_results(
  results = list("run" = list(
    "optimizer1" = list(
      "value" = 1, "comparable" = FALSE
    ),
    "optimizer2" = list(
      "value" = 2, "comparable" = TRUE
     )
  )),
  simplify = TRUE
)
}
}
\keyword{internal}
