% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{subset_argument}
\alias{subset_argument}
\title{Subset argument}
\usage{
subset_argument(argument, by_row, how, proportion, centers, ignore, seed)
}
\arguments{
\item{argument}{A \code{numeric} \code{vector}, \code{matrix}, or \code{data.frame}.}

\item{by_row, how, proportion, centers, ignore, seed}{See documentation of method \code{$reduce()} from \code{Nop} object.}
}
\value{
The standardized \code{argument}.
}
\description{
This helper function subsets an argument.
}
\examples{
\dontrun{
subset_argument(
  argument = diag(1:6), by_row = TRUE, how = "similar", proportion = 0.5,
  centers = 2, ignore = 1:2, seed = 1
)
}
}
\keyword{internal}
