% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategies.R
\name{subset_initialization}
\alias{subset_initialization}
\title{Subset initialization}
\usage{
subset_initialization(x, arg = "data", how = "random", prop = 0.5, row = TRUE)
}
\arguments{
\item{x}{An object of class \code{ino}.}

\item{arg}{A character, the name of the argument to be subsetted. The argument must be
of class \code{matrix} or \code{data.frame}.}

\item{how}{A character, specifying how to select the subset. Can be one of
\code{"random"} (default), \code{"first"}, and \code{"kmeans"}, see the
details.}

\item{prop}{A numeric between 0 and 1, specifying the proportion of the subset.
Can also be a vector of proportions.}

\item{row}{A boolean, set to \code{TRUE} to subset by row, set to \code{FALSE} to subset
by column.}
}
\value{
The updated input \code{x}.
}
\description{
This function is an implementation of the subset initialization strategy.
}
\details{
See the vignette on initialization strategies for more details.
}
\examples{
#data_subset_initialization()

}
\keyword{strategy}
