% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluations.R
\name{summary.ino}
\alias{summary.ino}
\title{Summary of initialization runs}
\usage{
\method{summary}{ino}(object, ...)
}
\arguments{
\item{object}{An object of class \code{ino}.}

\item{...}{Named function for computing statistics.
See function ... for an overview.}
}
\value{
A \code{tibble}, optimization runs as rows and variables as columns.
}
\description{
This function gives an overview of the initialization runs in an \code{ino}
object.
}
\details{
The following variables are available for each \code{ino} object:
\describe{
\item{.strategy}{the name of the initialization strategy}
\item{.time}{the optimization time}
\item{.optimum}{the function value at the optimum}
\item{.optimizer}{the identifier of the optimizer}
}
}
\keyword{evaluation}
