% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{test_ino}
\alias{test_ino}
\title{Test of an \code{ino} object}
\usage{
test_ino(x, verbose = FALSE)
}
\arguments{
\item{x}{An object of class \code{ino}.}

\item{verbose}{Set to \code{TRUE} to show the print the test results of the setup to the
console.}
}
\value{
The updated object \code{x} (invisibly).
}
\description{
This helper function tests the specification of an \code{ino} object.
}
\keyword{specification}
