% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ino.R
\name{update_opt}
\alias{update_opt}
\title{Update optimizer}
\usage{
update_opt(x, opt, verbose = getOption("ino_progress"))
}
\arguments{
\item{x}{An object of class \code{ino}.}

\item{opt}{The output of \code{\link{set_optimizer}}, which is an object of class
\code{optimizer}.
Per default, \code{opt = set_optimizer_nlm()}, which specifies the
\code{\link[stats]{nlm}} optimizer.
Can also be a (named) list of multiple \code{optimizer} objects.}

\item{verbose}{A boolean, which indicates whether progress should be printed.
Set to \code{TRUE} (\code{FALSE}) to print (hide) progress.
The default is \code{getOption("ino_progress")}, which is set to \code{TRUE}
when the package is loaded.}
}
\value{
An object of class \code{ino}.
}
\description{
Use this function to update the optimizer functions for an \code{ino} object.
}
\seealso{
\code{\link[=set_optimizer]{set_optimizer()}} to specify an optimizer.
}
\keyword{specification}
